### 2023-03-30 #5315 @kthyng
Hi @changliao1025 and thanks for your submission. Can you provide more information about the other paper submitted that uses this package? Specifically we want to be sure that 1. the other paper review did not entail a software review, since that is what JOSS does, and 2. there is substantial enough scholarly effort to justify two papers — usually this means a high standard for the software package (my initial quick on the repo check looked good!).
