### 2023-03-31 #5315 @changliao1025
The paper published recently is [Topological Relationship-Based Flow Direction Modeling: Mesh-Independent River Networks Representation](https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2022MS003089). The main objective of this paper is to provide scientific evidence and practical applications for this mesh-independent model. It demonstrates that the new method can achieve comparable or even superior results than the conventional methods.

Our JOSS submission highlights the quality and potential applications of this model for the hydrology community. We provide detailed documentation and case study examples to enable the use of our open-source model in various scenarios.
