### 2023-05-05 #5446 @editorialbot
[REVIEW]: PyFlowline a mesh independent river network generator for hydrologic models
**Submitting author:** @changliao1025 (<a href="http://orcid.org/0000-0002-7348-8858">Chang Liao</a>)
**Repository:** https://github.com/changliao1025/pyflowline
**Branch with paper.md** (empty if default branch): main
**Version:** v0.3.4
