### 2023-05-15 ecd379d
update vectorized function
