### 2023-05-17 #5446 @observingClouds
Hi @smchartrand, @andres-patrignani,
I just want to check if you have any questions regarding the review process. As a start, each of you should run `@editorialbot generate my checklist` to get a checklist of the individual tasks a JOSS review requires. Tasks can then be ticked off one by one. You can find additional information in [the guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Also, feel free to tag me here if you have any additional questions.

Cheers!
