### 2023-05-30 changliao1025/pyflowline#136 @smchartrand
Software License
This issue is related to the JOSS submission (https://github.com/openjournals/joss-reviews/issues/5446). Can the authors of pyflowline please reference an OSI approved license (https://opensource.org/licenses/) that is most similar to the one provided in the existing license.md, or if the existing was modified from an OSI approved license? Or perhaps Battelle Memorial Institute can submit their license for inclusion in the OSI approved licenses (other national labs have done so)? Thanks.
