### 2023-05-31 changliao1025/pyflowline#137 @smchartrand
Jupyter Notebook
The documentation under Quick Start points to a Jupyter Notebook named: mpas_notebook.ipynb. In the Notebooks directory the Jupyter Notebook is named: mpas_example.ipynb. I assume the _example is the correct file name to reference in the Quick Start.
