### 2023-05-31 #5446 @smchartrand
I want to start by congratulating @changliao1025 (and Matt Cooper) for their submission. The software authors have tackled a long standing problem in hydrologic modelling, and the outcome is impressive. Thanks for the opportunity to review your submission, and get to explore the code in more detail. I hope my comments are helpful.

Most of my comments below address either (a) challenges with using the .ipynb and .py codes provided in the repository for the Susquehanna example to generate results, or (b) setting the context of the code with respect to existing platforms. All the comments are relatively minor, or moderate in scope, and should be easily addressed. I organized my comments under the review checklist headers to simplify things.

### General Checks:

1. Repository: good.
2. License: I opened an issue in the home repository regarding the License.
3. Contribution and authorship, scholarly effort and data sharing:  good.
4. Reproducibility - The primary challenge I encountered was using the .ipynb file [in the notebooks directory] to generate results with pyflowline. The challenges are summarized in the following.

- Jupyter Notebook: I could not successfully plot the flow network using `oPyflowline.plot(sFilename_in = 'filter_flowline.png', sVariable_in = 'flowline_filter' )`. I did successfully plot the flow network using _geopandas_ (I added this to the notebook), and using QGIS. After lots of digging, I think the issue lies in the dependencies. The notebook requires cartopy>=0.21.0 [based on setup.py]; I run Ubuntu 22.04 LTS and python 3.7.10 [which is the maintained version for this OS distribution], with cartopy 0.18.0 as the standard install version. After many attempts and breaking my "apt", I attempted to use a virtual python environment and then decided to stop as the effort in to address the plotting issue was too large. Perhaps the authors can think about a different way to setup the notebook visualization, or make the required dependencies more clear without having to dig through the software repository to piece things together. _Geopandas_ was easy to implement in this case (_geopandas_ image of the raw network geometry copied and pasted here from mpas_example notebook), although some of the formatting may be lost or more difficult to implement.
![image](https://github.com/openjournals/joss-reviews/assets/20877310/31cde39c-6699-4b26-ad8f-76e740f81a88)

5. Human and animal reserach is n/a.

### Functionality:

1. Installation: good.
2. Functionality - I was able to follow the Quickstart and the Installation instructions separately to successfully install pyflowline. The "Functionality" needs some attention based on my comments above regarding use of the provided Jupyter Notebook mpas_example.ipynb [note: in the Quickstart the notebook is referred to by mpas_notebook.ipynb].
3. Performance: good - I found no performance issues and the example calculations for the Susquehanna ran in approximately 65 seconds on my laptop [ThinkPad X1 Carbon Intel i7 with 32 GB ram].

### Documentation:

1. Statement of need: good.
2. Installation instructions are good, _but a clear list of dependencies is missing_. There are three sources of dependencies and all three provide a different set: requirements_dev.txt [root directory], setup.py [root directory], and requirements.txt [docs directory]. Why are the listed dependencies different under each. Might it make sense to have a requirements.txt file in the root directory that can be linked to installation so the trace is more clear? Not sure if this would be the proper architecture so take the comment as a suggestion to have the dependencies clearly highlighted in one location.
3. Example usage: good - Authors provide a real-wrold example for the Susquehanna River.
4. Fucntionality documentation: good.
5. Automated tests are lacking, but use of the provided python script and Jupyter Notebook with the example offers a manual test of the software.
6. Community guidelines are clearly addressed in the License.

### Software Paper:

1. Summary - A clear statement of how the software benefits non-specialist audiences would be helpful.
2. Statement of need - More clear statement of who the intended or target audience is, would be helfpul. In the relating the current software to other and existing tools, the authors discuss just a few examples. I understand the contribution is a first of its kind, but it might be useful to expand this discussion a bit more. Is their merit in mentioning River Network Toolkit (http://rivtoolkit.com/)? Or RivGraph (https://joss.theoj.org/papers/10.21105/joss.02952)?
3. State of the field - Please elaborate a bit more on what existing packages can do, and why this is not enough (i.e. how pyflowline fills this gap). See comments above under Statement of Need.
4. References - See comment above under Statement of Need.
