### 2023-06-07 #5446 @changliao1025
@smchartrand Thank you for your comments, which will undoubtedly help us to provide better software.
We will address your concerns soon about some details in the document and example code.

Below are a few responses to your concerns:

- pyflowline was developed with visualization as an optional feature, see the setup file:

        extras_require={
           'visualization': ['cython', 'matplotlib', 'cartopy>=0.21.0']
        }

    This decision is made to reduce the model dependency. Users can run the model with only core dependency. But in the
    notebook, it is desired to have the visualization. We will update the documentation to clarify that, and users can use
    whatever methods for visualization, including QGIS, etc. And we recommend the Conda virtual environment for
    applications.

-  We will add some discussion comparing our model with others. Thanks for the suggestions for the other two similar tools. In short, our tool aims to provide the conceptual river networks for spatially-distributed hydrology models. Since these hydrologic models use the meshes to represent land surface, our model closes the gap by providing a method to define the river networks on top of meshes. This is different from vector line features.

We will address remaining comments in the coming weeks.
Thank you.
