+++ pyflowline/mesh/mpas/create_mpas_mesh.py
@@ -16,9 +16,11 @@ else:
 def create_mpas_mesh(iFlag_global_in, 
     iFlag_use_mesh_dem, 
     iFlag_save_mesh_in, 
-    pPolygon_in, 
+    
     sFilename_mesh_netcdf_in, 
-    sFilename_output_in): 
+    sFilename_output_in,
+    iFlag_antarctic_in=None,
+    pBoundary_in = None): 
     """
     Create a MPAS mesh
 
@@ -26,13 +28,23 @@ def create_mpas_mesh(iFlag_global_in,
         iFlag_global_in (int): _description_
         iFlag_use_mesh_dem (int): _description_
         iFlag_save_mesh_in (int): _description_
-        pPolygon_in (_type_): _description_
+        pBoundary_in (_type_): _description_
         sFilename_mesh_netcdf_in (_type_): _description_
         sFilename_output_in (_type_): _description_
 
     Returns:
         _type_: _description_
     """
+
+    if iFlag_antarctic_in is None:
+        iFlag_antarctic=0
+    else:
+        iFlag_antarctic=1
+
+    if pBoundary_in is None:
+        pBoundary = None
+    else:
+        pBoundary = pBoundary_in
        
     if (os.path.exists(sFilename_mesh_netcdf_in)):
         pass
@@ -42,6 +54,8 @@ def create_mpas_mesh(iFlag_global_in,
     
     if os.path.exists(sFilename_output_in):  
         os.remove(sFilename_output_in)
+
+    iFlag_remove_ice = 0
 
     pDatasets_in = Dataset(sFilename_mesh_netcdf_in)
 
@@ -186,12 +200,22 @@ def create_mpas_mesh(iFlag_global_in,
     for i in range(ncell):
         dLat = convert_360_to_180 (aLatitudeCell[i])
         dLon = convert_360_to_180 (aLongitudeCell[i])
-        pCenter = ogr.Geometry(ogr.wkbPoint)
-        pCenter.AddPoint(dLon, dLat)
-        pCenter1 = loads( pCenter.ExportToWkt() )
-        
-        
-        iFlag = pCenter1.within(pPolygon_in)
+
+        if iFlag_antarctic == 1:
+            #if it is antarctic, we dont need the boundary
+            if dLat < -60:
+                iFlag = True
+            else:  
+                iFlag = False
+            pass
+        else:
+            pCenter = ogr.Geometry(ogr.wkbPoint)
+            pCenter.AddPoint(dLon, dLat)
+            pCenter1 = loads( pCenter.ExportToWkt() )   
+            iFlag = pCenter1.within(pBoundary)
+            pass 
+
+
         if ( iFlag == True ):
             #get cell edge
             lCellID = int(aIndexToCellID[i])
@@ -199,76 +223,83 @@ def create_mpas_mesh(iFlag_global_in,
             dElevation_profile0 = float(aBed_elevation_profile[i,0])
             dThickness_ice = float( aIce_thickness[i] )
             dArea = float(aCellArea[i])
-            if dThickness_ice > 0:
-                continue
+            if iFlag_remove_ice == 1:
+                if dThickness_ice > 0 :
+                    continue
+                else:
+                    pass
             else:
-                aCellOnCellIndex = np.array(aCellsOnCell[i,:])
-                aEdgesOnCellIndex = np.array(aEdgesOnCell[i,:])
-                aVertexOnCellIndex = np.array(aVertexOnCell[i,:])
-                dummy0 = np.where(aVertexOnCellIndex > 0)
-                aVertexIndex = aVertexOnCellIndex[dummy0]
-                dummy1 = np.where(aEdgesOnCellIndex > 0)
-                aEdgeIndex= aEdgesOnCellIndex[dummy1]
-                dummy2 = np.where(aCellOnCellIndex > 0)
[…]
