### 2023-06-11 changliao1025/pyflowline#136 @changliao1025
The previous license is based on BSD 3-Clause "New" or "Revised" License but contains a format error. It has now been updated.
https://github.com/changliao1025/pyflowline/blob/main/LICENSE
