+++ pyflowline/classes/edge.py
@@ -11,8 +11,8 @@ if iFlag_cython is not None:
     from pyflowline.algorithms.cython.kernel import calculate_angle_betwen_vertex
     from pyflowline.algorithms.cython.kernel import calculate_distance_to_plane
 else:
-    from pyflowline.algorithms.auxiliary.gdal_functions import  calculate_angle_betwen_vertex
-    from pyflowline.algorithms.auxiliary.gdal_functions import calculate_distance_to_plane
+    from pyflowline.external.pyearth.gis.gdal.gdal_functions import  calculate_angle_betwen_vertex
+    from pyflowline.external.pyearth.gis.gdal.gdal_functions import calculate_distance_to_plane
 
 class EdgeClassEncoder(JSONEncoder):
     def default(self, obj):
