+++ pyflowline/external/pyearth/visual/map/map_multiple_vector_data.py
@@ -1,0 +1,376 @@
+import os
+import numpy as np
+
+import cartopy.crs as ccrs
+import cartopy.mpl.ticker as ticker
+import matplotlib as mpl
+from shapely.wkt import loads
+from osgeo import  osr, gdal, ogr
+
+
+import matplotlib.pyplot as plt
+import matplotlib.path as mpath
+import matplotlib.ticker as mticker
+import matplotlib.patches as mpatches
+import matplotlib.cm as cm
+
+
+from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
+
+from pyflowline.external.pyearth.toolbox.math.stat.remap import remap
+
+pProjection = ccrs.PlateCarree() #for latlon data only
+
+class OOMFormatter(mpl.ticker.ScalarFormatter):
+    def __init__(self, order=0, fformat="%1.1e", offset=True, mathText=True):
+        self.oom = order
+        self.fformat = fformat
+        mpl.ticker.ScalarFormatter.__init__(self,useOffset=offset,useMathText=mathText)
+    def _set_order_of_magnitude(self):
+        self.orderOfMagnitude = self.oom
+    def _set_format(self, vmin=None, vmax=None):
+        self.format = self.fformat
+        if self._useMathText:
+            self.format = r'$\mathdefault{%s}$' % self.format
+
+def map_multiple_vector_data(aFiletype_in,
+                             aFilename_in, 
+                             iFlag_colorbar_in=None,
+                        aFlag_thickness_in=None,
+                            aFlag_color_in=None,
+                            aVariable_in = None,
+                            sFilename_output_in=None,
+                            iFlag_scientific_notation_colorbar_in=None,
+                            sColormap_in = None,
+                            sTitle_in = None, 
+                            iDPI_in = None,
+                            dMissing_value_in=None,
+                            dData_max_in = None, 
+                            dData_min_in = None,
+                            sExtend_in =None,
+                            sFont_in = None,
+                            sUnit_in=None,
+                            aLegend_in = None,
+                            aExtent_in = None,
+                            pProjection_map_in=None):
+    """
+    plot vector data on a map
+    currently only support geojson and shapefile
+    by default, the program will plot all the polygons in the file
+    in the furture, the program will support to plot only a subset of polygons
+
+    Args:
+        iFiletype_in (_type_): _description_
+        sFilename_in (_type_): _description_
+        sFilename_output_in (_type_): _description_
+        iFlag_scientific_notation_colorbar_in (_type_, optional): _description_. Defaults to None.
+        sColormap_in (_type_, optional): _description_. Defaults to None.
+        sTitle_in (_type_, optional): _description_. Defaults to None.
+        iDPI_in (_type_, optional): _description_. Defaults to None.
+        dMissing_value_in (_type_, optional): _description_. Defaults to None.
+        dData_max_in (_type_, optional): _description_. Defaults to None.
+        dData_min_in (_type_, optional): _description_. Defaults to None.
+        sExtend_in (_type_, optional): _description_. Defaults to None.
+        sUnit_in (_type_, optional): _description_. Defaults to None.
+        aLegend_in (_type_, optional): _description_. Defaults to None.
+    """     
+    
+   
+    pDriver = ogr.GetDriverByName('GeoJSON')
+
+
[…]
