+++ pyflowline/external/pyearth/visual/map/map_multiple_vector_data.py
@@ -85,6 +85,11 @@ def map_multiple_vector_data(aFiletype_in,
     else:
         aFlag_thickness = aFlag_thickness_in
 
+    if aFlag_color_in is None:
+        aFlag_color= np.zeros(nFile, dtype=np.int16)
+    else:
+        aFlag_color = aFlag_color_in
+
     #get the extent first
     sFilename_in = aFilename_in[0]
 
@@ -214,29 +219,31 @@ def map_multiple_vector_data(aFiletype_in,
     iThickness_max = 2.5
     iThickness_min = 0.3
     for i in range(nFile):
-        sFilename = aFilename_in[i]
-       
+        sFilename = aFilename_in[i]       
         iFlag_thickness = aFlag_thickness[i]
+        iFlag_colar = aFlag_color[i]
         pDataset = pDriver.Open(sFilename, gdal.GA_ReadOnly)
         pLayer = pDataset.GetLayer(0)
 
-        if iFlag_thickness ==1 :  
-            sVariable = aVariable_in[i]
-            aValue = np.array(aValue_all[i])
-            if iFlag_data_min == 1  and iFlag_data_max ==1: #both are provided
-                aValue = np.clip(aValue, dData_min, dData_max)
-                dValue_max = dData_max # np.max(aValue)
-                dValue_min = dData_min # np.min(aValue)
-            else:
+        nColor = pLayer.GetFeatureCount()
+        aColor = cm.rainbow(np.linspace(0, 1, nColor))
+        lID = 0       
+        for pFeature in pLayer:
+            pGeometry_in = pFeature.GetGeometryRef()
+            sGeometry_type = pGeometry_in.GetGeometryName()
 
-                aValue = aValue[aValue != dMissing_value]
-                dValue_max = np.max(aValue)
-                dValue_min = np.min(aValue)
-                pass
-
-            for pFeature in pLayer:
-                pGeometry_in = pFeature.GetGeometryRef()
-                sGeometry_type = pGeometry_in.GetGeometryName()
+            if iFlag_thickness ==1 :  
+                sVariable = aVariable_in[i]
+                aValue = np.array(aValue_all[i])
+                if iFlag_data_min == 1  and iFlag_data_max ==1: #both are provided
+                    aValue = np.clip(aValue, dData_min, dData_max)
+                    dValue_max = dData_max 
+                    dValue_min = dData_min 
+                else:
+                    aValue = aValue[aValue != dMissing_value]
+                    dValue_max = np.max(aValue)
+                    dValue_min = np.min(aValue)
+                    pass
                 dValue = float(pFeature.GetField(sVariable))
 
                 if dValue != dMissing_value:
@@ -246,63 +253,63 @@ def map_multiple_vector_data(aFiletype_in,
                     if dValue < dValue_min:
                         dValue = dValue_min
 
-                iColor_index = int( (dValue - dValue_min) / (dValue_max - dValue_min) * 255 )
-                iThickness = remap( dValue, dValue_min, dValue_max, iThickness_min, iThickness_max )
+                iThickness = remap( dValue, dValue_min, dValue_max, iThickness_min, iThickness_max )    
+            else:
+                iThickness = 1.0
+            
+            if iFlag_colar ==1:
+                if nColor < 10:
+                    sColor = aColor[lID]
+                else:
+                    sVariable = aVariable_in[i]
+                    aValue = np.array(aValue_all[i])
+                    if iFlag_data_min == 1  and iFlag_data_max ==1: #both are provided
+                        aValue = np.clip(aValue, dData_min, dData_max)
+                        dValue_max = dData_max 
+                        dValue_min = dData_min 
+                    else:
[…]
