+++ pyflowline/classes/_visual.py
@@ -9,23 +9,23 @@ from pyflowline.external.pyearth.visual.map.map_vector_polygon_data import map_v
 from pyflowline.external.pyearth.visual.map.map_vector_polyline_data import map_vector_polyline_data
 from pyflowline.external.pyearth.visual.map.map_multiple_vector_data import map_multiple_vector_data
 
+#plot function
 def plot(self, sFilename_output_in=None,iFlag_title = None, sVariable_in=None, aExtent_in = None):
     if sVariable_in == 'mesh':
         self._plot_mesh(sFilename_output_in=sFilename_output_in, aExtent_in = aExtent_in)
-    else:            
+    else:
         if sVariable_in == 'overlap':
             self._plot_mesh_with_flowline( sFilename_output_in=sFilename_output_in, iFlag_title= iFlag_title, aExtent_in=aExtent_in)
-        else:            
-            for pBasin in self.aBasin:  
-                pBasin.basin_plot(self.iCase_index, 
-                                  self.iMesh_type, 
-                                  self.sMesh_type, 
-                                  sFilename_output_in=sFilename_output_in,
-                    iFlag_title= iFlag_title, 
-                    sVariable_in= sVariable_in, 
-                    aExtent_in=aExtent_in)
+        else:
+            for pBasin in self.aBasin:
+                pBasin.basin_plot(self.iCase_index,
+                                  self.iMesh_type,
+                                  self.sMesh_type,
+                                  sFilename_output_in=sFilename_output_in,                                 
+                                  sVariable_in= sVariable_in,
+                                  aExtent_in=aExtent_in)
             pass
-    
+
     return
 
 #this function is used to plot the study area
@@ -35,62 +35,72 @@ def _plot_study_area(self, sFilename_boundary_in = None, sFilename_slope_in = No
 def _plot_mesh(self, sFilename_output_in=None, aExtent_in=None, pProjection_map_in = None):
 
     sFilename_in = self.sFilename_mesh
+    sMesh_type = self.sMesh_type
 
-    map_vector_polygon_data(sFilename_in, sFilename_output_in=sFilename_output_in)
-    
-      
+    map_vector_polygon_data(sFilename_in, sFilename_output_in=sFilename_output_in, sTitle_in= sMesh_type)
+
+
     return
 
 
 #plot both polygon and polyline
-def _plot_mesh_with_flowline(self, sFilename_output_in=None, iFlag_title=None, aExtent_in=None, pProjection_map_in = None):
-    
+def _plot_mesh_with_flowline(self, 
+                             sFilename_output_in=None, 
+                             iFlag_title=None, 
+                             aExtent_in=None, 
+                             pProjection_map_in = None):
+
     aFiletype_in = list()
     aFilename_in = list()
     aFilename_in.append(self.sFilename_mesh)
     aFiletype_in.append(1)
 
-    for pBasin in self.aBasin: 
+    for pBasin in self.aBasin:
         aFiletype_in.append(2)
         dummy = pBasin.sFilename_flowline_conceptual
         sFilename_json = os.path.join(pBasin.sWorkspace_output_basin, dummy)
-                       
         aFilename_in.append(sFilename_json)
 
-    
-  
-    map_multiple_vector_data(aFiletype_in, aFilename_in, sFilename_output_in=sFilename_output_in)
+    map_multiple_vector_data(aFiletype_in, aFilename_in, 
+                             sFilename_output_in=sFilename_output_in, 
+                             sTitle_in= 'Mesh with flowline',
+                             aFlag_color_in=[0, 1])
     return
 
 #this is a reserved function
[…]
