+++ examples/susquehanna/run_simulation_mpas.py
@@ -31,7 +31,6 @@ iFlag_visualization = 1
 sMesh = 'mpas'
 sDate='20230101'
 
-
 oPyflowline = pyflowline_read_model_configuration_file(sFilename_configuration_in, 
     iCase_index_in=iCase_index, sDate_in=sDate)
 
@@ -41,27 +40,29 @@ oPyflowline.aBasin[0].dLongitude_outlet_degree=-76.009300
 
 oPyflowline.setup()
 if iFlag_visualization ==1:
-
     oPyflowline.plot(sFilename_output_in = 'filter_flowline.png', sVariable_in = 'flowline_filter' )
+    pass
 
-#oPyflowline.flowline_simplification()
+oPyflowline.flowline_simplification()
 
 if iFlag_visualization == 1:
      
     aExtent_meander = [-76.5,-76.2, 41.6,41.9] 
-    #oPyflowline.plot( iFlag_title=1 ,sVariable_in='flowline_simplified' , aExtent_in =aExtent_meander, sFilename_output_in = 'flowline_simplified.png', ) 
+    oPyflowline.plot( iFlag_title=1 ,sVariable_in='flowline_simplified' , sFilename_output_in = 'flowline_simplified.png', ) 
+
+    oPyflowline.plot( iFlag_title=1 ,sVariable_in='flowline_simplified' , aExtent_in =aExtent_meander, sFilename_output_in = 'flowline_simplified_zoom.png', ) 
 
     pass
-#aCell = oPyflowline.mesh_generation()
+aCell = oPyflowline.mesh_generation()
 
 if iFlag_visualization == 1:
-    #oPyflowline.plot( iFlag_title=1 ,sVariable_in='mesh', sFilename_output_in = 'mesh.png' ) 
+    oPyflowline.plot( iFlag_title=1 ,sVariable_in='mesh', sFilename_output_in = 'mesh.png' ) 
     pass
 
-#oPyflowline.reconstruct_topological_relationship(aCell)
+oPyflowline.reconstruct_topological_relationship(aCell)
 
 if iFlag_visualization == 1:
-    oPyflowline.plot( iFlag_title=1 ,sVariable_in='overlap', sFilename_output_in = 'mesh_w_flowline.png',)
+    oPyflowline.plot( iFlag_title=1, sVariable_in='overlap', sFilename_output_in = 'mesh_w_flowline.png',)
     pass
 oPyflowline.export()
 

[…]
