+++ paper.bib
@@ -36,3 +36,16 @@ eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2022MS003089},
 note = {e2022MS003089 2022MS003089},
 abstract = {Abstract River networks are important features in surface hydrology. However, accurately representing river networks in spatially distributed hydrologic and Earth system models is often sensitive to the model’s spatial resolution. Specifically, river networks are often misrepresented because of the mismatch between the model’s spatial resolution and river network details, resulting in significant uncertainty in the projected flow direction. In this study, we developed a topological relationship-based river network representation method for spatially distributed hydrologic models. This novel method uses (1) graph theory algorithms to simplify real-world vector-based river networks and assist in mesh generation; and (2) a topological relationship-based method to reconstruct conceptual river networks. The main advantages of our method are that (1) it combines the strengths of vector-based and DEM raster-based river network extraction methods; and (2) it is mesh-independent and can be applied to both structured and unstructured meshes. This method paves a path for advanced terrain analysis and hydrologic modeling across different scales.}
 }
+
+@article{Schwenk:2021, 
+doi = {10.21105/joss.02952}, 
+url = {https://doi.org/10.21105/joss.02952}, 
+year = {2021}, 
+publisher = {The Open Journal}, 
+volume = {6}, 
+number = {59}, 
+pages = {2952}, 
+author = {Jon Schwenk and Jayaram Hariharan}, 
+title = {RivGraph: Automatic extraction and analysis of river and delta channel network topology}, 
+journal = {Journal of Open Source Software} 
+}
