+++ paper.bib
@@ -49,3 +49,26 @@ author = {Jon Schwenk and Jayaram Hariharan},
 title = {RivGraph: Automatic extraction and analysis of river and delta channel network topology}, 
 journal = {Journal of Open Source Software} 
 }
+
+@techreport{Esri:2011,
+	title = {Arc {Hydro} {Tools} - {Tutorial} [{Software}]},
+	author = {{Esri Water Resources Team}},
+	year = {2011},
+}
+
+@article{Wu:2012,
+author = {Wu, Huan and Kimball, John S. and Li, Hongyi and Huang, Maoyi and Leung, L. Ruby and Adler, Robert F.},
+title = {A new global river network database for macroscale hydrologic modeling},
+journal = {Water Resources Research},
+volume = {48},
+number = {9},
+pages = {},
+keywords = {upscaling, DRT, river network, hydrography},
+doi = {https://doi.org/10.1029/2012WR012313},
+url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2012WR012313},
+eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2012WR012313},
+abstract = {Coarse-resolution (upscaled) river networks are critical inputs for runoff routing in macroscale hydrologic models. Recently, Wu et al. (2011) developed a hierarchical dominant river tracing (DRT) algorithm for automated extraction and spatial upscaling of river networks using fine-scale hydrography inputs. We applied the DRT algorithms using combined HydroSHEDS and HYDRO1k global fine-scale hydrography inputs and produced a new series of upscaled global river network data at multiple (1/16° to 2°) spatial resolutions. The new upscaled results are internally consistent and congruent with the baseline fine-scale inputs and should facilitate improved regional to global scale hydrologic simulations.},
+year = {2012}
+}
+
+
