+++ paper.bib
@@ -1,6 +1,7 @@
 @article{Engwirda:2021,
   title={‘UNIFIED’LAGUERRE-POWER MESHES FOR COUPLED EARTH SYSTEM MODELLING},
   author={Engwirda, Darren and Liao, Chang},
+  year = {2021},
   doi = {10.5281/zenodo.5558988},
 }
 
@@ -26,6 +27,7 @@
 author = {Liao, Chang and Zhou, Tian and Xu, Donghui and Cooper, Matthew G and Engwirda, Darren and Li, Hong-Yi and Leung, L. Ruby},
 title = {Topological relationship-based flow direction modeling: Mesh-independent river networks representation},
 journal = {Journal of Advances in Modeling Earth Systems},
+year = {2022},
 volume = {n/a},
 number = {n/a},
 pages = {e2022MS003089},
