### 2023-06-14 001964f
add installation test
