+++ paper.bib
@@ -73,4 +73,24 @@ abstract = {Coarse-resolution (upscaled) river networks are critical inputs for 
 year = {2012}
 }
 
+@software{LiaoPyearth:2022,
+  author       = {Chang Liao},
+  title        = {{PyEarth: A lightweight Python package for Earth 
+                   science}},
+  month        = mar,
+  year         = 2022,
+  publisher    = {Zenodo},
+  version      = {0.1.20},
+  doi          = {10.5281/zenodo.6368652},
+  url          = {https://doi.org/10.5281/zenodo.6368652}
+}
+
+@article{Sahr:2011,
+  title={User documentation for discrete global grid generation software},
+  author={Sahr, K},
+  journal={Southern Oregon Univ., Ashland, OR, USA, Tech. Rep. Dggrid version},
+  volume={3},
+  year={2011}
+}
+
 
