+++ pyflowline/classes/pycase.py
@@ -1,12 +1,18 @@
 import os
+import stat
 from pathlib import Path
 import json
 from json import JSONEncoder
 import datetime
 import importlib
+import subprocess
+from shutil import copy2
 import numpy as np
 from osgeo import ogr, osr
 from shapely.wkt import loads
+
+from pyflowline.external.pyearth.system.define_global_variables import *
+
 from pyflowline.classes.mpas import pympas
 from pyflowline.classes.hexagon import pyhexagon
 from pyflowline.classes.latlon import pylatlon
@@ -25,6 +31,7 @@ from pyflowline.mesh.hexagon.create_hexagon_mesh import create_hexagon_mesh
 from pyflowline.mesh.latlon.create_latlon_mesh import create_latlon_mesh
 from pyflowline.mesh.square.create_square_mesh import create_square_mesh
 from pyflowline.mesh.mpas.create_mpas_mesh import create_mpas_mesh
+from pyflowline.mesh.dggrid.create_dggrid_mesh import create_dggrid_mesh
 from pyflowline.mesh.tin.create_tin_mesh import create_tin_mesh
 
 pDate = datetime.datetime.today()
@@ -39,11 +46,11 @@ class CaseClassEncoder(JSONEncoder):
         if isinstance(obj, np.ndarray):
             return obj.tolist()
         if isinstance(obj, list):
-            pass  
+            pass
         if isinstance(obj, pyvertex):
             return json.loads(obj.tojson()) #lVertexID
         if isinstance(obj, pyedge):
-            return obj.lEdgeID        
+            return obj.lEdgeID
         if isinstance(obj, pyflowline):
             return obj.lFlowlineID
         if isinstance(obj, pyhexagon):
@@ -53,10 +60,10 @@ class CaseClassEncoder(JSONEncoder):
         if isinstance(obj, pylatlon):
             return obj.lCellID
         if isinstance(obj, pympas):
-            return obj.lCellID        
+            return obj.lCellID
         if isinstance(obj, pybasin):
-            return obj.lBasinID    
-            
+            return obj.lBasinID
+
         return JSONEncoder.default(self, obj)
 
 class flowlinecase(object):
@@ -70,7 +77,7 @@ class flowlinecase(object):
         flowlinecase: A flowlinecase object
     """
     iCase_index= 0
-    
+
     sMesh_type = 1
     iFlag_standalone=1
     iFlag_flowline = 1
@@ -85,12 +92,13 @@ class flowlinecase(object):
     iFlag_create_mesh=1
     iFlag_intersect = 1
     iFlag_rotation=0
-    iFlag_break_by_distance = 0  #if the distance between two vertice are far, 
+    iFlag_break_by_distance = 0  #if the distance between two vertice are far,
+    iFlag_dggrid = 0
     nOutlet = 1 #by default , there shoule ne only one ouelet
     dResolution_degree=0.0
     dResolution_meter=0.0
     dThreshold_small_river=0.0
-    dThreshold_break_by_distance = 5000.0 
+    dThreshold_break_by_distance = 5000.0
     dLongitude_left = -180
     dLongitude_right = 180
     dLatitude_bot = -90
@@ -98,30 +106,32 @@ class flowlinecase(object):
     dElevation_mean=-9999.0
     sFilename_model_configuration=''
     sFilename_mesh_boundary = ''
-    sWorkspace_input=''      
-    sWorkspace_output=''    
-    #sWorkspace_output_case=''    
+    sWorkspace_input=''
+    sWorkspace_output=''
+    #sWorkspace_output_case=''
     sRegion=''
     sModel=''
     sMesh_type ='hexagon'
 
[…]
