+++ pyflowline/mesh/dggrid/create_dggrid_mesh.py
@@ -2,176 +2,324 @@
 import sys, os, stat
 import numpy as np
 from pathlib import Path
-from shutil import copy2
 import subprocess
 import datetime
+from shutil import copy2
+from osgeo import osr, ogr, gdal
+from shapely.wkt import loads
+from pyflowline.formats.convert_coordinates import convert_gcs_coordinates_to_cell
 
 from pyflowline.external.pyearth.system.define_global_variables import *
+from pyflowline.formats.convert_attributes import convert_gcs_attributes_to_cell
 
 pDate = datetime.datetime.today()
 sDate_default = "{:04d}".format(pDate.year) + "{:02d}".format(pDate.month) + "{:02d}".format(pDate.day)
 
-
-def create_dggrid_mesh( iCase_index, 
-    iResolution_index ,
-        sGrid_type,
-            sDate_in = None, 
-        sFilename_crop_shapefile_in =None ):
-
-    sCase = "{:03d}".format( iCase_index )
-    sResolution = "{:02d}".format( iResolution_index )
-
-    if sDate_in is not None:
-        sDate = sDate_in 
+def generate_bash_script(sWorkspace_output):
+    sName  = 'dggrid.ini'
+    sFilename_configuration  =  os.path.join(sWorkspace_output,  sName )
+    os.chdir(sWorkspace_output)
+    #detemine the system platform
+    # Determine the appropriate executable name for the platform
+    system = platform.system()
+    if platform.system() == 'Windows':
+        sFilename_executable = 'dggrid.exe'
+        iFlag_unix = 0
     else:
-        sDate = sDate_default
+        sFilename_executable = './dggrid'
 
-    if sFilename_crop_shapefile_in is not None:
-        if os.path.isfile(sFilename_crop_shapefile_in):
+    if system == 'Windows':
+        # execute binary on Windows
+        iFlag_unix = 0
+    elif system == 'Linux':
+        # execute binary on Linux
+        iFlag_unix = 1
+    elif system == 'Darwin':
+        # execute binary on macOS
+        iFlag_unix = 1
+    else:
+        # unsupported operating system
+        print('Unsupported operating system: ' + system)
+        print('Please reach out to the developers for assistance.')
+        #generate the bash/batch script
+    if iFlag_unix == 1 :
+        sFilename_bash = os.path.join(str(Path(  sWorkspace_output)  ) ,  "run_dggrid.sh" )
+        ofs = open(sFilename_bash, 'w')
+        sLine = '#!/bin/bash\n'
+        ofs.write(sLine)
+        sLine = 'cd ' +   sWorkspace_output+ '\n'
+        ofs.write(sLine)
+        sLine = sFilename_executable + ' ' + sFilename_configuration + '\n'
+        ofs.write(sLine)
+        ofs.close()
+        os.chmod(sFilename_bash, stat.S_IRWXU )
+    else:
+        sFilename_bash = os.path.join(str(Path(  sWorkspace_output)  ) ,  "run_dggrid.bat" )
+        ofs = open(sFilename_bash, 'w')
+        sLine = 'cd ' +   sWorkspace_output+ '\n'
+        ofs.write(sLine)
+        sLine = sFilename_executable + ' ' + sFilename_configuration + '\n'
+        ofs.write(sLine)
+        ofs.close()
+        os.chmod(sFilename_bash, stat.S_IRWXU )
+
+    return
+
+def find_number_range(number, aArray):
[…]
