+++ pyflowline/pyflowline_read_model_configuration_file.py
@@ -110,6 +110,7 @@ def pyflowline_read_model_configuration_file(sFilename_configuration_in,
 
     aConfig['sDate'] = sDate
     aConfig['sModel'] = sModel
+    aConfig['sMesh_type'] = sMesh_type
     aConfig['sWorkspace_output'] = sWorkspace_output
 
     aConfig["sFilename_model_configuration"] = sFilename_configuration_in
