+++ pyflowline/classes/vertex.py
@@ -8,7 +8,7 @@ iFlag_cython = importlib.util.find_spec("cython")
 if iFlag_cython is not None:
     from pyflowline.algorithms.cython.kernel import calculate_distance_based_on_lon_lat
 else:
-    from pyflowline.algorithms.auxiliary.gdal_functions import calculate_distance_based_on_lon_lat
+    from pyflowline.external.pyearth.gis.gdal.gdal_functions import calculate_distance_based_on_lon_lat
 
 class VertexClassEncoder(JSONEncoder):
     def default(self, obj):
