### 2023-07-07 39c14b3
fix the hole and neighbor issue
