+++ pyflowline/mesh/mpas/create_mpas_mesh copy.py
@@ -15,7 +15,8 @@ else:
 
 def create_mpas_mesh(iFlag_global_in, 
     iFlag_use_mesh_dem, 
-    iFlag_save_mesh_in,     
+    iFlag_save_mesh_in, 
+    
     sFilename_mesh_netcdf_in, 
     sFilename_output_in,
     iFlag_antarctic_in=None,
@@ -200,8 +201,7 @@ def create_mpas_mesh(iFlag_global_in,
     ncell = len(aIndexToCellID) 
     aMpas = list()
 
-    #add a mpas cell into a list
-    def add_cell_into_list(aList, i, lCellID, dArea,dElevation_mean,dElevation_profile0, aCoords  ):
+    def add_cell_into_list(aList, i):
         dLat = convert_360_to_180 (aLatitudeCell[i])
         dLon = convert_360_to_180 (aLongitudeCell[i])
         #vertex
@@ -215,7 +215,9 @@ def create_mpas_mesh(iFlag_global_in,
         dummy2 = np.where(aCellOnCellIndex > 0)
         aNeighborIndex= (aCellOnCellIndex[dummy2]).astype(int)
         aVertexIndexOnEdge = np.array(aVertexOnEdge0[aEdgeIndex-1,:]).astype((int))
-    
+        aLonVertex = aLongitudeVertex[aVertexIndex-1]
+        aLatVertex = aLatitudeVertex[aVertexIndex-1]
+        nVertex = len(aLonVertex)
         pmpas = convert_gcs_attributes_to_cell(4, dLon, dLat, aCoords, aVertexIndex, aEdgeIndex, aVertexIndexOnEdge)               
         pmpas.dArea = dArea
         pmpas.calculate_edge_length()
@@ -223,32 +225,19 @@ def create_mpas_mesh(iFlag_global_in,
         pmpas.lCellID = lCellID
         pmpas.dElevation_mean  = dElevation_mean
         pmpas.dElevation_profile0 = dElevation_profile0
-        #now setup the neighbor information
         pmpas.aNeighbor=aNeighborIndex
         pmpas.nNeighbor=len(aNeighborIndex)
-        if pmpas.nNeighbor != pmpas.nVertex:
-            #this cell is next to the ocean boundary
-            pmpas.nNeighbor_land = pmpas.nNeighbor 
-            pmpas.nNeighbor_ocean = pmpas.nVertex - pmpas.nNeighbor
-            pmpas.aNeighbor_land=aNeighborIndex
-            pmpas.nNeighbor_land=len(aNeighborIndex)
-            print(lCellID)
-            print('Warning: nNeighbor != nVertex')
-        else:
-            pmpas.nNeighbor_land = pmpas.nNeighbor 
-            pmpas.nNeighbor_ocean = 0
-            pmpas.aNeighbor_land = aNeighborIndex        
-            pmpas.nNeighbor_land=len(aNeighborIndex)
-
+        pmpas.aNeighbor_land=aNeighborIndex
+        pmpas.nNeighbor_land=len(aNeighborIndex)
         aDistance=list()
         for i in range(pmpas.nNeighbor):
+            lNeighborID = pmpas.aNeighbor[i]
             #find shared edge
             lEdgeID= aEdgeIndex[i]              
             lIndex = lEdgeID-1
             dDistance = aDcEdge[lIndex]
             aDistance.append(dDistance)
             pass
-
         pmpas.aNeighbor_distance = aDistance
         aList.append(pmpas)
         return aList
@@ -268,14 +257,24 @@ def create_mpas_mesh(iFlag_global_in,
             #center
             dLat = convert_360_to_180 (aLatitudeCell[i])
[…]
