+++ pyflowline/mesh/mpas/create_mpas_mesh copy.py
@@ -1,481 +1,0 @@
-import os
-import math
-import importlib
-import numpy as np
-from osgeo import ogr, osr, gdal
-from netCDF4 import Dataset
-from shapely.wkt import loads
-from pyflowline.formats.convert_attributes import convert_gcs_attributes_to_cell
-gdal.UseExceptions()  
-iFlag_cython = importlib.util.find_spec("cython") 
-if iFlag_cython is not None:
-    from pyflowline.algorithms.cython.kernel import convert_360_to_180
-else:
-    from pyflowline.external.pyearth.gis.gdal.gdal_functions import convert_360_to_180
-
-def create_mpas_mesh(iFlag_global_in, 
-    iFlag_use_mesh_dem, 
-    iFlag_save_mesh_in, 
-    
-    sFilename_mesh_netcdf_in, 
-    sFilename_output_in,
-    iFlag_antarctic_in=None,
-    pBoundary_in = None): 
-    """
-    Create a MPAS mesh
-
-    Args:
-        iFlag_global_in (int): _description_
-        iFlag_use_mesh_dem (int): _description_
-        iFlag_save_mesh_in (int): _description_
-        pBoundary_in (_type_): _description_
-        sFilename_mesh_netcdf_in (_type_): _description_
-        sFilename_output_in (_type_): _description_
-
-    Returns:
-        _type_: _description_
-    """
-
-    if iFlag_antarctic_in is None:
-        iFlag_antarctic=0
-    else:
-        iFlag_antarctic=iFlag_antarctic_in
-
-    if pBoundary_in is None:
-        pBoundary = None
-    else:
-        #for the reason that a geometry object will be crash if the associated dataset is closed, we must pass wkt string
-        #https://gdal.org/api/python_gotchas.html
-        pBoundary = ogr.CreateGeometryFromWkt(pBoundary_in)
-        #pBoundary = pBoundary_in #only works for shapely geometry object
-       
-    if (os.path.exists(sFilename_mesh_netcdf_in)):
-        pass
-    else:
-        print('Mesh file does not exist!')
-        return
-    
-    if os.path.exists(sFilename_output_in):  
-        os.remove(sFilename_output_in)
-
-    
-
-    pDatasets_in = Dataset(sFilename_mesh_netcdf_in)
-
-    netcdf_format = pDatasets_in.file_format
-    pDriver_geojson = ogr.GetDriverByName('GeoJSON')     
-    pSpatial_reference_gcs = osr.SpatialReference()  
-    pSpatial_reference_gcs.ImportFromEPSG(4326)    # WGS84 lat/lon    
-    #geojson
-    if iFlag_save_mesh_in ==1:
-        pDataset = pDriver_geojson.CreateDataSource(sFilename_output_in)
-        pLayer = pDataset.CreateLayer('cell', pSpatial_reference_gcs, ogr.wkbPolygon)
-        # Add one attribute
-        pLayer.CreateField(ogr.FieldDefn('id', ogr.OFTInteger64)) #long type for high resolution
-        pLayer.CreateField(ogr.FieldDefn('lon', ogr.OFTReal)) #long type for high resolution
-        pLayer.CreateField(ogr.FieldDefn('lat', ogr.OFTReal)) #long type for high resolution
-        pArea_field = ogr.FieldDefn('area', ogr.OFTReal)
-        pArea_field.SetWidth(20)
-        pArea_field.SetPrecision(2)
-        pLayer.CreateField(pArea_field)
-        if iFlag_use_mesh_dem == 1:
-            pLayer.CreateField(ogr.FieldDefn('elev', ogr.OFTReal)) #float type for high resolution
[…]
