+++ pyflowline/external/pyearth/visual/map/map_vector_polyline_data.py
@@ -1,17 +1,14 @@
 import os
 import numpy as np
 from osgeo import  osr, gdal, ogr
-
+from shapely.wkt import loads
 import matplotlib.pyplot as plt
-from matplotlib import cm
+import matplotlib.cm as cm
 import matplotlib as mpl
 import matplotlib.ticker as mticker
 import matplotlib.path as mpath
-from shapely.wkt import loads
 import cartopy.crs as ccrs
-import cartopy.mpl.ticker as ticker
 from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
-
 from pyflowline.external.pyearth.toolbox.math.stat.remap import remap
 
 class OOMFormatter(mpl.ticker.ScalarFormatter):
@@ -31,8 +28,8 @@ def map_vector_polyline_data(sFilename_in,
                              iFlag_color_in = None,
                              iFlag_label_in = None,
                              iFlag_thickness_in =None,
-                             sField_thickness_in=None,       
-                             sField_color_in=None,                     
+                             sField_thickness_in=None,
+                             sField_color_in=None,
                              iFlag_scientific_notation_colorbar_in=None,
                              sColormap_in = None,
                              sTitle_in = None,
@@ -97,9 +94,9 @@ def map_vector_polyline_data(sFilename_in,
         sField_thickness = sField_thickness_in
     else:
         sField_thickness = ''
-    
+
     pDriver = ogr.GetDriverByName('GeoJSON')
-    
+
     pDataset = pDriver.Open(sFilename_in, gdal.GA_ReadOnly)
     pLayer = pDataset.GetLayer(0)
 
@@ -125,16 +122,16 @@ def map_vector_polyline_data(sFilename_in,
         sUnit = sUnit_in
     else:
         sUnit =  ''
-    
+
     if sFont_in is not None:
         sFont = sFont_in
-    else:    
+    else:
         sFont = "Times New Roman"
 
     plt.rcParams["font.family"] = sFont
-    
-    cmap = cm.get_cmap(sColormap)  
- 
+
+    cmap = cm.get_cmap(sColormap)
+
     pSrs = osr.SpatialReference()
     pSrs.ImportFromEPSG(4326)    # WGS84 lat/lon
 
@@ -146,7 +143,7 @@ def map_vector_polyline_data(sFilename_in,
     for pFeature in pLayer:
         pGeometry_in = pFeature.GetGeometryRef()
         sGeometry_type = pGeometry_in.GetGeometryName()
-        
+
         if sGeometry_type =='LINESTRING':
             dummy0 = loads( pGeometry_in.ExportToWkt() )
             aCoords_gcs = dummy0.coords
@@ -156,14 +153,16 @@ def map_vector_polyline_data(sFilename_in,
             dLon_max = np.max( [dLon_max, np.max(aCoords_gcs[:,0])] )
             dLon_min = np.min( [dLon_min, np.min(aCoords_gcs[:,0])] )
             dLat_max = np.max( [dLat_max, np.max(aCoords_gcs[:,1])] )
-            dLat_min = np.min( [dLat_min, np.min(aCoords_gcs[:,1])] )     
-                    
+            dLat_min = np.min( [dLat_min, np.min(aCoords_gcs[:,1])] )
+
     if pProjection_map_in is not None:
         pProjection_map = pProjection_map_in
     else:
-        pProjection_map = ccrs.Orthographic(central_longitude =  0.50*(dLon_max+dLon_min),  central_latitude = 0.50*(dLat_max+dLat_min), globe=None)
-   
-   
+        pProjection_map = ccrs.Orthographic(central_longitude =  0.50*(dLon_max+dLon_min),
[…]

[…]
