+++ pyflowline/classes/_visual.py
@@ -7,59 +7,81 @@ from pyflowline.external.pyearth.visual.map.map_vector_polyline_data import map_
 from pyflowline.external.pyearth.visual.map.map_multiple_vector_data import map_multiple_vector_data
 
 #plot function
-def plot(self, 
-        iFlag_type_in = None,
-         iFlag_title = None, 
-        sFilename_output_in=None,
-        sVariable_in=None, 
-        aExtent_in = None,
-        pProjection_map_in = None):
-    if iFlag_type_in == 1: #polygon based 
-        if sVariable_in == 'mesh':
-            self._plot_mesh(sFilename_output_in=sFilename_output_in, 
-                            aExtent_in = aExtent_in, 
-                            pProjection_map_in = pProjection_map_in)
+def plot(self,
+         iFlag_type_in = None,
+         iFlag_title_in = None,
+         sFilename_output_in=None,
+         sVariable_in=None,
+         aExtent_in = None,
+         pProjection_map_in = None):
+    """_summary_
+
+    Args:
+        iFlag_type_in (_type_, optional): _description_. Defaults to None.
+        iFlag_title_in (_type_, optional): _description_. Defaults to None.
+        sFilename_output_in (_type_, optional): _description_. Defaults to None.
+        sVariable_in (_type_, optional): _description_. Defaults to None.
+        aExtent_in (_type_, optional): _description_. Defaults to None.
+        pProjection_map_in (_type_, optional): _description_. Defaults to None.
+    """
+    
+
+    if iFlag_type_in == 1: #point based, such as dam
+        #currently, this feature is not supported
+        pass
     else:
-        if iFlag_type_in == 2: #polyline based, only flowline
-            #this one is similar to the basin_plot, may be added later
-            pass
+        if iFlag_type_in == 2:
+            #polyline based, only flowline
+            for pBasin in self.aBasin:
+                pBasin.basin_plot(self.iCase_index,
+                                  self.iMesh_type,
+                                  self.sMesh_type,
+                                  iFlag_title_in= iFlag_title_in,
+                                  sFilename_output_in=sFilename_output_in,
+                                  sVariable_in= sVariable_in,
+                                  aExtent_in=aExtent_in,
+                                  pProjection_map_in = pProjection_map_in)
         else:
-            if iFlag_type_in == 3: #mixed based
-                if sVariable_in == 'overlap':
-                    self._plot_mesh_with_flowline( sFilename_output_in=sFilename_output_in, 
-                                                  iFlag_title= iFlag_title, 
-                                                  aExtent_in=aExtent_in, 
-                                                  pProjection_map_in = pProjection_map_in)
-                else:
-                    for pBasin in self.aBasin:
-                        pBasin.basin_plot(self.iCase_index,
-                                          self.iMesh_type,
-                                          self.sMesh_type,
-                                          sFilename_output_in=sFilename_output_in,                                 
-                                          sVariable_in= sVariable_in,
-                                          aExtent_in=aExtent_in,
-                                           pProjection_map_in = pProjection_map_in)
+            if iFlag_type_in == 3: #polygon based
+                if sVariable_in == 'mesh':
+                    self._plot_mesh(sFilename_output_in=sFilename_output_in,
+                                    aExtent_in = aExtent_in,
+                                    pProjection_map_in = pProjection_map_in)
 
-                pass
             else:
-                pass
+                if iFlag_type_in == 4: #mixed based mesh + flowline
+                    if sVariable_in == 'overlap':
+                        self._plot_mesh_with_flowline(sFilename_output_in=sFilename_output_in,
+                                                      iFlag_title_in= iFlag_title_in,
+                                                      aExtent_in=aExtent_in,
[…]
