+++ pyflowline/classes/basin.py
@@ -117,12 +117,22 @@ class pybasin(object):
     sFilename_flowline_conceptual_info=''
     sFilename_confluence_simplified_info=''
     sFilename_confluence_conceptual_info=''    
+
     aFlowline_basin_filtered=None
     aFlowline_basin_simplified=None
     aFlowline_basin_conceptual=None    
     pVertex_outlet=None
     aConfluence_basin_simplified= None
     aConfluence_basin_conceptual= None
+
+
+    #for hexwatershed compatibility
+    sFilename_elevation=''
+    sFilename_slope=''
+    sFilename_drainage_area=''
+    sFilename_flow_direction ='' 
+    sFilename_distance_to_outlet = ''
+
 
     
     iFlag_visual = importlib.util.find_spec("cartopy") 
@@ -194,12 +204,12 @@ class pybasin(object):
         if 'sFilename_flowline_filter' in aParameter:
             self.sFilename_flowline_filter = aParameter['sFilename_flowline_filter']
         else:
-            self.sFilename_flowline_filter   = ''
+            self.sFilename_flowline_filter = ''
 
         if 'sWorkspace_output_basin' in aParameter:
             self.sWorkspace_output_basin = aParameter['sWorkspace_output_basin']
         else:
-            self.sWorkspace_output_basin   = '.'
+            self.sWorkspace_output_basin = '.'
             
         Path(self.sWorkspace_output_basin).mkdir(parents=True, exist_ok=True)
 
@@ -208,15 +218,14 @@ class pybasin(object):
         if 'sFilename_dam' in aParameter:
             self.sFilename_dam = aParameter['sFilename_dam']
         else:
-            self.sFilename_dam   = ''
+            self.sFilename_dam  = ''
 
         if 'sFilename_flowline_topo' in aParameter:
             self.sFilename_flowline_topo = aParameter['sFilename_flowline_topo']
         else:
-            self.sFilename_flowline_topo   =''
+            self.sFilename_flowline_topo = ''
 
         self.sBasinID  = "{:08d}".format(self.lBasinID)
-
 
         if not os.path.isfile(self.sFilename_flowline_filter):
             print("The filtered flowline file does not exist!")
@@ -230,20 +239,30 @@ class pybasin(object):
                 exit
             pass
 
-        #geojson
-        self.sFilename_flowline_segment_index_before_intersect = 'flowline_segment_index_before_intersect.geojson'
-        self.sFilename_flowline_simplified = 'flowline_simplified.geojson'
-        self.sFilename_flowline_split = 'flowline_split.geojson'
-        self.sFilename_flowline_intersect  = 'flowline_intersect_mesh.geojson'
-        self.sFilename_flowline_conceptual = 'flowline_conceptual.geojson'
-        self.sFilename_flowline_edge = 'flowline_edge.geojson'
-        self.sFilename_area_of_difference = 'area_of_difference.geojson'
+        #geojson, full path of the file
+        self.sFilename_flowline_segment_index_before_intersect = os.path.join(str(self.sWorkspace_output_basin ),'flowline_segment_index_before_intersect.geojson')
+        self.sFilename_flowline_simplified = os.path.join(str(self.sWorkspace_output_basin ),'flowline_simplified.geojson')
+        self.sFilename_flowline_split = os.path.join(str(self.sWorkspace_output_basin ),'flowline_split.geojson')
+        self.sFilename_flowline_intersect  = os.path.join(str(self.sWorkspace_output_basin ),'flowline_intersect_mesh.geojson')
+        self.sFilename_flowline_conceptual = os.path.join(str(self.sWorkspace_output_basin ),'flowline_conceptual.geojson')
[…]
