+++ pyflowline/classes/basin.py
@@ -19,8 +19,6 @@ from pyflowline.formats.export_flowline import export_flowline_to_geojson
 from pyflowline.formats.export_vertex import export_vertex_to_geojson
 from pyflowline.external.pyearth.toolbox.reader.text_reader_string import text_reader_string
 
-
-
 from pyflowline.algorithms.split.find_flowline_vertex import find_flowline_vertex
 from pyflowline.algorithms.split.find_flowline_confluence import find_flowline_confluence
 from pyflowline.algorithms.split.split_flowline import split_flowline
@@ -125,14 +123,18 @@ class pybasin(object):
     aConfluence_basin_simplified= None
     aConfluence_basin_conceptual= None
 
+    #json
+    sFilename_hexwatershed_json=''
+    sFilename_stream_edge_json =''
 
-    #for hexwatershed compatibility
+    #geojson for hexwatershed compatibility
     sFilename_elevation=''
     sFilename_slope=''
     sFilename_drainage_area=''
     sFilename_flow_direction ='' 
     sFilename_distance_to_outlet = ''
-
+    sFilename_stream_segment=''
+    sFilename_stream_edge=''
 
     
     iFlag_visual = importlib.util.find_spec("cartopy") 
@@ -239,7 +241,18 @@ class pybasin(object):
                 exit
             pass
 
+        #json
+        self.sFilename_hexwatershed_json = os.path.join(str(self.sWorkspace_output_basin ), "watershed.json" )
+        self.sFilename_stream_edge_json = os.path.join(str(self.sWorkspace_output_basin ), 'stream_edge.json')
+        self.sFilename_basin_info = os.path.join(str(self.sWorkspace_output_basin ), 'basin_info.json')
+        self.sFilename_flowline_conceptual_info = os.path.join(str(self.sWorkspace_output_basin ), 'flowline_conceptual_info.json')
+        self.sFilename_flowline_simplified_info = os.path.join(str(self.sWorkspace_output_basin ), 'flowline_simplified_info.json')
+        self.sFilename_confluence_conceptual_info = os.path.join(str(self.sWorkspace_output_basin ),'confluence_conceptual_info.json')
+        self.sFilename_confluence_simplified_info = os.path.join(str(self.sWorkspace_output_basin ),'confluence_simplified_info.json')
+        
         #geojson, full path of the file
+        #full paths are required for the following files
+        #for hexwatershed compatibility, the geojson files will be generated by the pyhexwatershed front end
         self.sFilename_flowline_segment_index_before_intersect = os.path.join(str(self.sWorkspace_output_basin ),'flowline_segment_index_before_intersect.geojson')
         self.sFilename_flowline_simplified = os.path.join(str(self.sWorkspace_output_basin ),'flowline_simplified.geojson')
         self.sFilename_flowline_split = os.path.join(str(self.sWorkspace_output_basin ),'flowline_split.geojson')
@@ -247,21 +260,14 @@ class pybasin(object):
         self.sFilename_flowline_conceptual = os.path.join(str(self.sWorkspace_output_basin ),'flowline_conceptual.geojson')
         self.sFilename_flowline_edge = os.path.join(str(self.sWorkspace_output_basin ),'flowline_edge.geojson')
         self.sFilename_area_of_difference = os.path.join(str(self.sWorkspace_output_basin ),'area_of_difference.geojson')
-
-        self.sFilename_basin_info = os.path.join(str(self.sWorkspace_output_basin ), 'basin_info.json')
-        self.sFilename_flowline_conceptual_info = os.path.join(str(self.sWorkspace_output_basin ), 'flowline_conceptual_info.json')
[…]
