+++ pyflowline/external/pyearth/visual/animate/animate_vector_polygon_data.py
@@ -6,7 +6,7 @@ import cartopy.crs as ccrs
 import cartopy.mpl.ticker as ticker
 import matplotlib as mpl
 from shapely.wkt import loads
-from osgeo import  osr, gdal, ogr
+from osgeo import osr, gdal, ogr
 
 
 import matplotlib.pyplot as plt
@@ -14,176 +14,223 @@ import matplotlib.path as mpath
 import matplotlib.ticker as mticker
 import matplotlib.patches as mpatches
 import matplotlib.cm as cm
+from matplotlib import animation
 from pyearth.toolbox.data.cgpercentiles import cgpercentiles
 from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
 
-pProjection = ccrs.PlateCarree() #for latlon data only
+pProjection = ccrs.PlateCarree()  # for latlon data only
+
+
 class OOMFormatter(mpl.ticker.ScalarFormatter):
     def __init__(self, order=0, fformat="%1.1e", offset=True, mathText=True):
         self.oom = order
         self.fformat = fformat
-        mpl.ticker.ScalarFormatter.__init__(self,useOffset=offset,useMathText=mathText)
+        mpl.ticker.ScalarFormatter.__init__(
+            self, useOffset=offset, useMathText=mathText
+        )
+
     def _set_order_of_magnitude(self):
         self.orderOfMagnitude = self.oom
+
     def _set_format(self, vmin=None, vmax=None):
         self.format = self.fformat
         if self._useMathText:
-            self.format = r'$\mathdefault{%s}$' % self.format
+            self.format = r"$\mathdefault{%s}$" % self.format
+
 
 iFigwidth_default = 12
 iFigheight_default = 12
 
-def animate_vector_polygon_data(sFilename_mesh_in, 
-                                sFilename_animation_json_in,
-        iFlag_type_in = None, 
-        iFigwidth_in=None, 
-        iFigheight_in=None,
-            aExtent_in = None,
-        pProjection_map_in = None):
 
+def animate_vector_polygon_data(
+    sFilename_mesh_in,
+    sFilename_animation_json_in,
+    sFilename_animation_out,
+    iFlag_type_in=None,
+    iFigwidth_in=None,
+    iFigheight_in=None,
+    aExtent_in=None,
+    pProjection_map_in=None,
+):
     if iFigwidth_in is None:
         iFigwidth_in = iFigwidth_default
 
     if iFigheight_in is None:
-        iFigheight_in = iFigheight_default 
+        iFigheight_in = iFigheight_default
 
+    # read domain json
+    # read result json
 
-    #read domain json 
-    #read result json
-   
-    #read animation json
-   
-    
+    # read animation json
+
     dLat_min = 90
     dLat_max = -90
     dLon_min = 180
-    dLon_max = -180 
-    aCellID = list() 
+    dLon_max = -180
+    aCellID = list()
     aData = list()
     aCell_raw = list()
     aCell_new = list()
     aCell_animation = list()
-    pDriver = ogr.GetDriverByName('GeoJSON')
-   
+    pDriver = ogr.GetDriverByName("GeoJSON")
+
     pDataset = pDriver.Open(sFilename_mesh_in, gdal.GA_ReadOnly)
     pLayer = pDataset.GetLayer(0)
-    pSrs = osr.SpatialReference()  
-    pSrs.ImportFromEPSG(4326)    # WGS84 lat/lon
+    pSrs = osr.SpatialReference()
+    pSrs.ImportFromEPSG(4326)  # WGS84 lat/lon
[…]
