### 2023-07-20 #5446 @changliao1025
Hi, @smchartrand ,
Thank you for taking the time to write a review of our work. I appreciate your feedback. Below is the full response to your comments.
Thank you for reviewing our submission and recognizing our modeling tool. PyFlowline was developed to close several gaps in hydrologic modeling. For example, although it is designed to run at regional and global scales using any mesh type, many modelers may be more interested in the watershed scale using structured meshes. To this extent, most existing applications can be viewed as a special case in PyFlowline, and PyFlowline provides an opportunity to expand existing applications to other domains and meshes.

**General Checks reply:**

- Thanks for the comment.

- There was a format issue in our original license file. It has now been resolved. The current license is [BSD-3-Clause license](https://github.com/changliao1025/pyflowline/blob/main/LICENSE).

- Thanks for the comment.

- We understand that there needs to be some clarification regarding the visualization component of the workflow. We have improved the visualization component in both the source code and documentation.

1. First, we revised the visualization class and function completely. Because PyFlowline is a core component in HexWatershed and both models share similar visualization patterns (point, polyline, polygon, and mixed), we merged all the visualization feature into an external module within PyFlowline. This module is largely based on another Python package PyEarth, developed by @chango1025. We plan to add PyEarth as a dependency package.

2. Although visualization is an important component in PyFlowline, it is not a required step to run the model. Besides, our choice of using GeoJSON as the main data I/O format is to ease the visualization task as many tools can visualize GeoJSON. Given the complexity of structured/unstructured datasets that contain point/polyline/polygon, we list the visualization feature as experimental, so users should expect undesired behavior.

3. We added Geopandas as another option in the notebook so users can easily switch to Geopandas for a quick fix.

4. Indeed, it is highly possible that the cartopy installation may affect the visualization. As discussed above, all the vector objects in PyFlowline are plotted using a geodetic framework. If some of the spatial reference information is missing, the vector objects may not be plotted correctly.

**Functionality reply:**

- Thanks for the comment.

- We updated the documentation, so there are consistent now.

- Susquehanna River basin is average size watershed with a moderate amount of mesh cells. If running with a large domain with more than 10k mesh cells and 100+ river channels, the cython feature may be used to improve the performance. Besides, our workflow can be run using a SLURM job on a high-performance computer to obtain the best performance.

**Documentation reply:**

- Thanks for the comment.

- We update the readme file to explicitly list the dependency. We also specifically separate the optional dependency. The readthedocs documentation is also updated to maintain consistency.

- Thanks for the comment.

- Thanks for the comment.

- Thanks for the comment. In most cases, our PyFlowline model should be run as a whole workflow. Running individual steps may not be feasible because the model requires all the objects to be accessible during simulation. In the future, we plan to add the `checkpoint` feature so the model can resume at certain points, which may support more unit test capability.

- Thanks for the comment.

**Software Paper reply:**

- Thanks for the comment. We revised the summary to highlight that the model output from our package can be used in general hydrologic models, including
`With PyFlowline, hydrologic modelers can generate conceptual river networks at various spatial resolutions for both structured and unstructured computational meshes. The generated river network datasets can be used by hydrologic models across scales.`

- Thanks for the comment. We added some background related to the existing vector-based river network method and explained why we developed this new model. We also cited related references. `For hydrologic modelers, river networks are a key input for hydrologic models. While some hydrologic models accept vector-based river networks [@Schwenk:2021], others only accept mesh cell-based, which require a generation method from the vector-based river network. Currently, generating a mesh cell-based river network from a given vector-based river network and arbitrary computational mesh is a major challenge. Existing methods are typically limited to structured rectangular meshes, such as 30m x 30m cartesian grids for high-resolution watershed-scale modeling or 0.5 degree x 0.5 degree geographic grids for global climate modeling.`

- Thanks for the comment. We provided details of several different methods:

Existing river network representation methods often fall into these three categories:

1. Vector-based, hydrologic models that use this method cannot couple river and land because there is no one-to-one mapping [@Schwenk:2021];
2. High-resolution DEM-based, only supports structured rectangle grids (e.g., 30m x 30m ) at high spatial resolutions [@Esri:2011];
3. Upscaling-based, only supports structured geographic grids (e.g., 0.5 degree x 0.5 degree) at coarse resolutions [@Wu:2012]. This method often cannot provide global coverage, including Greenland and the Antarctic.

PyFlowline is the only modeling software that provides these unique features:

1. It can generate river networks on unstructured meshes;
2. It uses topological relationships to capture river networks precisely;
3. It can be applied at both high and coarse resolutions;
4. It can provide global coverage, including Greenland and the Antarctic.

Thank you.
