### 2023-07-21 eb17cfe
fix a bug for neighbors
