### 2023-07-22 changliao1025/pyflowline#164 @andres-patrignani
Read-only file system error
Receiving error: `OSError: [Errno 30] Read-only file system: '/compyfs'` when running the following line in the Jupyter Notebook example:

`oPyflowline = pyflowline_read_model_configuration_file(sFilename_configuration_in, iCase_index_in=iCase_index, sMesh_type_in= sMesh_type, sDate_in=sDate)`

Library loads fine and code cells before this one return the correct output. Not sure whether this is an issue with the library, a file configuration in the example folder, or something in my local system.
