### 2023-07-22 changliao1025/pyflowline#165 @andres-patrignani
Installation using conda
I'm having trouble installing the library using the following conda command from the documentation:

`conda install -c conda-forge pyflowline`

Here is the output:

```
(base) AGR-1011A-APAT1:soilwaterweb andrespatrignani$ conda install -c conda-forge pyflowline
Collecting package metadata (current_repodata.json): done
Solving environment: -
The environment is inconsistent, please check the package plan carefully
The following packages are causing the inconsistency:

  - defaults/osx-64::scrapy==2.8.0=py310hecd8cb5_0
  - defaults/osx-64::sphinx==5.0.2=py310hecd8cb5_0
  - defaults/osx-64::scikit-image==0.19.3=py310hcec6c5f_1
  - defaults/osx-64::parsel==1.6.0=py310hecd8cb5_0
  - defaults/osx-64::imbalanced-learn==0.10.1=py310hecd8cb5_0
  - defaults/osx-64::anaconda-client==1.11.1=py310hecd8cb5_0
  - defaults/noarch::requests-file==1.5.1=pyhd3eb1b0_0
  - defaults/osx-64::twisted==22.2.0=py310hca72f7f_1
  - defaults/osx-64::conda-repo-cli==1.0.27=py310hecd8cb5_0
  - defaults/noarch::asttokens==2.0.5=pyhd3eb1b0_0
  - defaults/noarch::cookiecutter==1.7.3=pyhd3eb1b0_0
  - defaults/noarch::itemloaders==1.0.4=pyhd3eb1b0_1
  - defaults/osx-64::holoviews==1.15.4=py310hecd8cb5_0
  - defaults/osx-64::bcrypt==3.2.0=py310hca72f7f_1
  - defaults/osx-64::nbclient==0.5.13=py310hecd8cb5_0
  - defaults/osx-64::jupyter==1.0.0=py310hecd8cb5_8
  - defaults/osx-64::spyder==5.4.1=py310hecd8cb5_0
  - defaults/osx-64::navigator-updater==0.3.0=py310hecd8cb5_0
  - defaults/osx-64::notebook==6.5.2=py310hecd8cb5_0
  - defaults/osx-64::colorcet==3.0.1=py310hecd8cb5_0
  - defaults/osx-64::pyct==0.5.0=py310hecd8cb5_0
  - defaults/osx-64::transformers==4.24.0=py310hecd8cb5_0
  - defaults/noarch::binaryornot==0.4.4=pyhd3eb1b0_1
  - defaults/osx-64::conda==23.1.0=py310hecd8cb5_0
  - defaults/osx-64::distributed==2022.7.0=py310hecd8cb5_0
  - defaults/osx-64::intake==0.6.7=py310hecd8cb5_0
  - defaults/noarch::pyhamcrest==2.0.2=pyhd3eb1b0_2
  - defaults/osx-64::multipledispatch==0.6.0=py310hecd8cb5_0
  - defaults/osx-64::ipython==8.10.0=py310hecd8cb5_0
  - defaults/osx-64::seaborn==0.12.2=py310hecd8cb5_0
  - defaults/osx-64::arrow==1.2.3=py310hecd8cb5_1
  - defaults/osx-64::pandas==1.5.3=py310h3ea8b11_0
  - defaults/osx-64::numpydoc==1.5.0=py310hecd8cb5_0
  - defaults/osx-64::spyder-kernels==2.4.1=py310hecd8cb5_0
  - defaults/noarch::python-dateutil==2.8.2=pyhd3eb1b0_0
  - defaults/osx-64::patsy==0.5.3=py310hecd8cb5_0
  - defaults/osx-64::huggingface_hub==0.10.1=py310hecd8cb5_0
  - defaults/noarch::conda-verify==3.4.2=py_1
  - defaults/osx-64::nbclassic==0.5.2=py310hecd8cb5_0
  - defaults/osx-64::qtconsole==5.4.0=py310hecd8cb5_0
  - defaults/osx-64::requests==2.28.1=py310hecd8cb5_0
  - defaults/osx-64::scikit-learn==1.2.1=py310hcec6c5f_0
  - defaults/noarch::protego==0.1.16=py_0
  - defaults/osx-64::anaconda-project==0.11.1=py310hecd8cb5_0
  - defaults/noarch::jupyterlab_pygments==0.1.2=py_0
  - defaults/noarch::jinja2-time==0.2.0=pyhd3eb1b0_3
  - defaults/osx-64::panel==0.14.3=py310hecd8cb5_0
  - defaults/osx-64::xarray==2022.11.0=py310hecd8cb5_0
  - defaults/noarch::w3lib==1.21.0=pyhd3eb1b0_0
  - defaults/osx-64::anaconda-navigator==2.4.0=py310hecd8cb5_0
  - defaults/noarch::pooch==1.4.0=pyhd3eb1b0_0
  - defaults/noarch::bleach==4.1.0=pyhd3eb1b0_0
  - defaults/osx-64::datashader==0.14.4=py310hecd8cb5_0
  - defaults/osx-64::widgetsnbextension==3.5.2=py310hecd8cb5_0
  - defaults/osx-64::scipy==1.10.0=py310ha516a68_1
  - defaults/osx-64::jupyterlab_server==2.19.0=py310hecd8cb5_0
  - defaults/osx-64::hvplot==0.8.2=py310hecd8cb5_0
  - defaults/osx-64::notebook-shim==0.2.2=py310hecd8cb5_0
  - defaults/noarch::ipywidgets==7.6.5=pyhd3eb1b0_1
  - defaults/osx-64::conda-build==3.23.3=py310hecd8cb5_0
  - defaults/osx-64::jupyter_server==1.23.4=py310hecd8cb5_0
  - defaults/osx-64::urllib3==1.26.14=py310hecd8cb5_0
  - defaults/osx-64::jupyter_console==6.6.2=py310hecd8cb5_0
  - defaults/osx-64::jupyterlab==3.5.3=py310hecd8cb5_0
  - defaults/osx-64::gensim==4.3.0=py310h3ea8b11_0
  - defaults/osx-64::matplotlib==3.7.0=py310hecd8cb5_0
  - defaults/noarch::automat==20.2.0=py_0
  - defaults/osx-64::statsmodels==0.13.5=py310h7b7cdfe_1
  - defaults/osx-64::ipykernel==6.19.2=py310h20db666_0
  - defaults/noarch::conda-token==0.4.0=pyhd3eb1b0_0
  - defaults/osx-64::jupyter_client==7.3.4=py310hecd8cb5_0
  - defaults/noarch::service_identity==18.1.0=pyhd3eb1b0_1
  - defaults/noarch::stack_data==0.2.0=pyhd3eb1b0_0
  - defaults/osx-64::websocket-client==0.58.0=py310hecd8cb5_4
  - defaults/osx-64::datashape==0.5.4=py310hecd8cb5_1
  - defaults/osx-64::dask==2022.7.0=py310hecd8cb5_0
  - defaults/osx-64::matplotlib-base==3.7.0=py310ha533b9c_0
  - defaults/noarch::tldextract==3.2.0=pyhd3eb1b0_0
  - defaults/osx-64::conda-content-trust==0.1.3=py310hecd8cb5_0
  - defaults/osx-64::nbconvert==6.5.4=py310hecd8cb5_0
done

==> WARNING: A newer version of conda exists. <==
  current version: 23.1.0
  latest version: 23.5.2

Please update conda by running

    $ conda update -n base -c defaults conda

Or to minimize the number of packages updated during conda update use

     conda install conda=23.5.2

## Package Plan ##

  environment location: /Users/andrespatrignani/anaconda3

  added / updated specs:
    - pyflowline

The following packages will be downloaded:

    package                    |            build
    ---------------------------|-----------------
    cryptography-38.0.4        |  py310hdd0c95c_0         1.0 MB  conda-forge
    curl-7.88.1                |       h6df9250_1         140 KB  conda-forge
    gdal-3.6.2                 |  py310h4d00ebc_2         1.4 MB
    geos-3.8.0                 |       hb1e8313_0         815 KB
    gettext-0.21.0             |       he85b6c0_1         3.0 MB
    hdf4-4.2.13                |       h39711bb_2         682 KB
    hdf5-1.12.1                |nompi_h0aa1fa2_104         3.7 MB  conda-forge
    kealib-1.5.0               |       h72c422f_1         151 KB
    krb5-1.20.1                |       h049b76e_0         1.1 MB  conda-forge
    libarchive-3.6.2           |       h29ab7a1_2         782 KB
    libboost-1.73.0            |      h3fa6bed_12        12.7 MB
    libcurl-7.88.1             |       h6df9250_1         336 KB  conda-forge
    libgdal-3.6.2              |       hd905d60_2         9.0 MB
    libkml-1.3.0               |       h85bf17e_6         466 KB
    libnetcdf-4.8.1            |       h9f5a9a2_3         1.1 MB
    libnghttp2-1.52.0          |       he2ab024_0         599 KB  conda-forge
    libpq-12.15                |       h04015c4_1         2.5 MB
    libspatialite-4.3.0a       |      hd1e8275_23         3.0 MB
    libsqlite-3.42.0           |       h58db7d2_0         858 KB  conda-forge
    libssh2-1.11.0             |       hd019ec5_0         253 KB  conda-forge
    libxml2-2.10.3             |       h930c0e2_0         665 KB
    libxslt-1.1.37             |       h5d22bc9_0         236 KB  conda-forge
    libzip-1.9.2               |       h6db710c_1         117 KB  conda-forge
    libzlib-1.2.13             |       h8a1eda9_5          58 KB  conda-forge
    lxml-4.9.2                 |  py310h6c40b1e_0         1.4 MB
    netcdf4-1.6.0              |nompi_py310hc1a5e9b_100         448 KB  conda-forge
    openssl-3.1.1              |       h8a1eda9_1         2.2 MB  conda-forge
    pcre2-10.42                |       h9b97e30_0         1.3 MB
    python-3.10.12             |had23ca6_0_cpython        12.5 MB  conda-forge
    python_abi-3.10            |          3_cp310           6 KB  conda-forge
    qhull-2020.2               |       h940c156_2         1.8 MB  conda-forge
    qtwebkit-5.212             |       hbfab81c_5        10.6 MB
    sqlite-3.42.0              |       h2b0dec6_0         854 KB  conda-forge
    tiledb-2.3.3               |       h1b93210_3         2.5 MB
    xz-5.4.2                   |       h6c40b1e_0         372 KB
    zlib-1.2.13                |       h8a1eda9_5          89 KB  conda-forge
    ------------------------------------------------------------
                                           Total:        78.5 MB

The following NEW packages will be INSTALLED:

  beautifulsoup4     conda-forge/noarch::beautifulsoup4-4.12.2-pyha770c72_0
  cairo              conda-forge/osx-64::cairo-1.16.0-h0ab9d94_1001
  cftime             conda-forge/osx-64::cftime-1.6.2-py310h936d966_1
  chardet            conda-forge/osx-64::chardet-5.1.0-py310h2ec42d9_0
  expat              conda-forge/osx-64::expat-2.5.0-hf0c8a7f_1
  fontconfig         conda-forge/osx-64::fontconfig-2.13.94-h10f422b_0
  freexl             conda-forge/osx-64::freexl-1.0.6-hb7f2c08_1
  gdal               pkgs/main/osx-64::gdal-3.6.2-py310h4d00ebc_2
  geos               pkgs/main/osx-64::geos-3.8.0-hb1e8313_0
  geotiff            pkgs/main/osx-64::geotiff-1.7.0-hf6ff7a4_1
  hdf4               pkgs/main/osx-64::hdf4-4.2.13-h39711bb_2
  json-c             conda-forge/osx-64::json-c-0.16-h01d06f9_0
  kealib             pkgs/main/osx-64::kealib-1.5.0-h72c422f_1
  libblas            conda-forge/osx-64::libblas-3.9.0-16_osx64_openblas
  libboost           pkgs/main/osx-64::libboost-1.73.0-h3fa6bed_12
  libcblas           conda-forge/osx-64::libcblas-3.9.0-16_osx64_openblas
  libexpat           conda-forge/osx-64::libexpat-2.5.0-hf0c8a7f_1
  libgdal            pkgs/main/osx-64::libgdal-3.6.2-hd905d60_2
  libkml             pkgs/main/osx-64::libkml-1.3.0-h85bf17e_6
  libnetcdf          pkgs/main/osx-64::libnetcdf-4.8.1-h9f5a9a2_3
  libspatialite      pkgs/main/osx-64::libspatialite-4.3.0a-hd1e8275_23
  libsqlite          conda-forge/osx-64::libsqlite-3.42.0-h58db7d2_0
  libzip             conda-forge/osx-64::libzip-1.9.2-h6db710c_1
  libzlib            conda-forge/osx-64::libzlib-1.2.13-h8a1eda9_5
  mkl                conda-forge/osx-64::mkl-2022.2.1-h44ed08c_16952
  netcdf4            conda-forge/osx-64::netcdf4-1.6.0-nompi_py310hc1a5e9b_100
  pcre2              pkgs/main/osx-64::pcre2-10.42-h9b97e30_0
  pixman             conda-forge/osx-64::pixman-0.38.0-h01d97ff_1003
  poppler            pkgs/main/osx-64::poppler-0.81.0-h031bf0f_2
  poppler-data       conda-forge/noarch::poppler-data-0.4.12-hd8ed1ab_0
  proj               conda-forge/osx-64::proj-6.2.1-h773a61f_0
  pyflowline         conda-forge/noarch::pyflowline-0.2.3-pyhd8ed1ab_0
  pygments           conda-forge/noarch::pygments-2.15.1-pyhd8ed1ab_0
  pyopenssl          conda-forge/noarch::pyopenssl-23.2.0-pyhd8ed1ab_1
  python_abi         conda-forge/osx-64::python_abi-3.10-3_cp310
  pytorch            conda-forge/osx-64::pytorch-1.12.1-cpu_py310h32957a8_0
  qhull              conda-forge/osx-64::qhull-2020.2-h940c156_2
  six                conda-forge/noarch::six-1.16.0-pyh6c4a22f_0
  sleef              conda-forge/osx-64::sleef-3.5.1-h6db0672_2
  tiledb             pkgs/main/osx-64::tiledb-2.3.3-h1b93210_3
  xerces-c           pkgs/main/osx-64::xerces-c-3.2.4-hd2b157f_0

The following packages will be REMOVED:

  h5py-3.7.0-py310h6c517f8_0
  pycurl-7.45.1-py310h0a4fc7d_0
  pytables-3.7.0-py310h59775c6_1

The following packages will be UPDATED:

  ca-certificates    pkgs/main::ca-certificates-2023.01.10~ --> conda-forge::ca-certificates-2023.5.7-h8857fd0_0
  certifi            pkgs/main/osx-64::certifi-2022.12.7-p~ --> conda-forge/noarch::certifi-2023.5.7-pyhd8ed1ab_0
  curl                    pkgs/main::curl-7.87.0-h6c40b1e_0 --> conda-forge::curl-7.88.1-h6df9250_1
  gettext                                 0.21.0-h7535e17_0 --> 0.21.0-he85b6c0_1
  hdf5                    pkgs/main::hdf5-1.10.6-h10fe05b_1 --> conda-forge::hdf5-1.12.1-nompi_h0aa1fa2_104
  krb5                    pkgs/main::krb5-1.19.4-hdba6334_0 --> conda-forge::krb5-1.20.1-h049b76e_0
  libarchive                               3.6.2-h65c5294_0 --> 3.6.2-h29ab7a1_2
  libcurl              pkgs/main::libcurl-7.87.0-ha585b31_0 --> conda-forge::libcurl-7.88.1-h6df9250_1
  libnghttp2         pkgs/main::libnghttp2-1.46.0-ha29bfda~ --> conda-forge::libnghttp2-1.52.0-he2ab024_0
  libpq                                     12.9-h1c9f633_3 --> 12.15-h04015c4_1
  libssh2              pkgs/main::libssh2-1.10.0-h0a4fc7d_0 --> conda-forge::libssh2-1.11.0-hd019ec5_0
  libxml2                                 2.9.14-hbf8cd5e_0 --> 2.10.3-h930c0e2_0
  libxslt              pkgs/main::libxslt-1.1.35-h5b33f42_0 --> conda-forge::libxslt-1.1.37-h5d22bc9_0
  lxml                                4.9.1-py310h65b224f_0 --> 4.9.2-py310h6c40b1e_0
  openssl              pkgs/main::openssl-1.1.1t-hca72f7f_0 --> conda-forge::openssl-3.1.1-h8a1eda9_1
  python                pkgs/main::python-3.10.9-h218abb5_1 --> conda-forge::python-3.10.12-had23ca6_0_cpython
  qtwebkit                                 5.212-h24dc246_4 --> 5.212-hbfab81c_5
  sqlite                pkgs/main::sqlite-3.40.1-h880c91c_0 --> conda-forge::sqlite-3.42.0-h2b0dec6_0
  xz                                      5.2.10-h6c40b1e_1 --> 5.4.2-h6c40b1e_0
  zlib                    pkgs/main::zlib-1.2.13-h4dc903c_0 --> conda-forge::zlib-1.2.13-h8a1eda9_5

The following packages will be SUPERSEDED by a higher-priority channel:

  cryptography       pkgs/main::cryptography-39.0.1-py310h~ --> conda-forge::cryptography-38.0.4-py310hdd0c95c_0

Proceed ([y]/n)? y

Downloading and Extracting Packages

Preparing transaction: done
Verifying transaction: failed

RemoveError: 'pyopenssl' is a dependency of conda and cannot be removed from
conda's operating environment.
```
