### 2023-07-22 a32e4d6
rename class member variable
