### 2023-07-22 changliao1025/pyflowline#165 @changliao1025
Hi, @andres-patrignani , have you tried to create a new conda environment before installing pyflowline?

    conda create --name pyflowline_test
    conda install -c conda-forge pyflowline

The error suggested an issue with the pyopenssl package, which is not used by pyflowline.
