### 2023-07-22 changliao1025/pyflowline#164 @changliao1025
Yes, the example provided in the repo was tested on our system, so you should change the paths to the actual path on your system. Hope that helps.
