+++ pyflowline/classes/_visual.py
@@ -1,10 +1,12 @@
 import os
+from pathlib import Path
 #dependency packages
 from pyflowline.external.pyearth.visual.map.map_vector_polygon_data import map_vector_polygon_data
 from pyflowline.external.pyearth.visual.map.map_vector_polyline_data import map_vector_polyline_data
 from pyflowline.external.pyearth.visual.map.map_multiple_vector_data import map_multiple_vector_data
 
 #plot function
+
 def plot(self,
          iFlag_type_in = None,
          iFlag_title_in = None,
@@ -46,8 +48,7 @@ def plot(self,
         if iFlag_type_in == 2:
             #polyline based, only flowline
             for pBasin in self.aBasin:
-                pBasin.basin_plot(self.iCase_index,
-                                  self.iMesh_type,
+                pBasin.basin_plot(self.iMesh_type,
                                   self.sMesh_type,
                                   iFlag_title_in= iFlag_title_in,
                                   sFilename_output_in=sFilename_output_in,
@@ -77,7 +78,10 @@ def plot(self,
 
         return
 
-def _plot_mesh(self, sFilename_output_in=None, aExtent_in=None, pProjection_map_in = None):
+def _plot_mesh(self, 
+               sFilename_output_in=None, 
+               aExtent_in=None, 
+               pProjection_map_in = None):
 
     sFilename_in = self.sFilename_mesh 
     sMesh_type = self.sMesh_type
@@ -120,90 +124,41 @@ def _plot_mesh_with_flowline(self,
                              pProjection_map_in = pProjection_map_in)
     return
 
-def basin_plot(self,
-               iCase_index,
-               iMesh_type,
-               sMesh_type,
-               sFilename_output_in=None,
-               iFlag_title_in=None,
-               sVariable_in=None,
-               aExtent_in = None,
-               pProjection_map_in = None):
-    """_summary_
-
-    Args:
-        iCase_index (_type_): _description_
-        iMesh_type (_type_): _description_
-        sMesh_type (_type_): _description_
-        sFilename_output_in (_type_, optional): _description_. Defaults to None.
-        iFlag_title_in (_type_, optional): _description_. Defaults to None.
-        sVariable_in (_type_, optional): _description_. Defaults to None.
-        aExtent_in (_type_, optional): _description_. Defaults to None.
-        pProjection_map_in (_type_, optional): _description_. Defaults to None.
-    """
-
-
-    iFlag_label = 0
-    sWorkspace_output_basin = self.sWorkspace_output_basin
-    if sVariable_in is not None:
-        if sVariable_in == 'flowline_raw':
-            sFilename_json = self.sFilename_flowline_raw
-            sTitle = 'Original flowline'
-            iFlag_color = 0
-        else:
-            if sVariable_in == 'flowline_filter':
-                sFilename_json = self.sFilename_flowline_filter
-                sTitle = 'Filtered flowline'
-                iFlag_color = 0
-            else:
-                if sVariable_in == 'flowline_simplified':
-                    sFilename_json = self.sFilename_flowline_simplified                    
-                    sTitle = 'Simplified flowline'
-                    iFlag_color = 1
-                    if aExtent_in is None:
-                        iFlag_label = 1
-                    else:
-                        iFlag_label=0
-                else:
-                    if sVariable_in == 'flowline_conceptual':
-                        sFilename_json = self.sFilename_flowline_conceptual                        
-                        sTitle = 'Conceptual flowline'
-                        iFlag_color = 1
-                        if aExtent_in is None:
-                            iFlag_label = 1
-                        else:
-                            iFlag_label=0
-                    else:
-                        if sVariable_in == 'aof':
-                            sFilename_json = self.sFilename_area_of_difference                            
[…]
