### 2023-07-23 changliao1025/pyflowline#165 @andres-patrignani
Thanks. This worked. I was able to install the module and dependencies. It may be good to add this to the documentation since people that have a working `conda` environment may not attempt to create a new one.
