### 2023-07-23 changliao1025/pyflowline#164 @andres-patrignani
It helps and is in the notebook instructions. Editing all the paths by hand in the `*.json` files seems a bit overwhelming for a simple example, unless I'm missing something. Thanks for the prompt help.
