### 2023-07-24 #5446 @andres-patrignani
### General checks

1. Repository: Good
2. License: License is available and easy to find.
3. Contribution and authorship: Author has made substantial contribution to the project repository. Paper authors re appropriate and complete.

### Functionality
1. Installation: Does installation proceed as outlined in the documentation? Yes, but it may be good to indicate that in some cases creating a new `conda` environment is necessary. It seems to me that most users having already a working `conda` environment will not attempt to do this.
2. Functionality: Good
3. Performance: I did not have any performance issue.

### Documentation

1. Overall the documentation is extensive with tables, figures, an “Overview” section stating the need for the package, and additional detailed information about how to use and get started with the package. I do have some comments to improve the language, which sometimes can lead to unclear statements that may result confusing to users. Below are some suggestions for improvement:

-	Consider adding the URL to the documentation page in both the “About” section of the repository (so that the link appears below “A mesh independent river network generator for hydrologic models”)

-	Consider moving the “Quickstart” section in the Readme file to the top, so that people can easily find and access the documentation page.

-	Consider the following re-worded sentences: “River networks are landscape features typically represented using vector layers. However, most hydrologic models rely on regular grids to discretize the spatial domain and cannot directly ingest vector features into the model. As a result, hydrologic models usually implement a so-called stream-burning process to convert the vector-based river network into a mesh-based river network. PyFlowline solves this issue by using a mesh-independent approach that intersects the vector river network and mesh to reconstruct the conceptual river network.”

-	The following sentence is offers little detail about limitations: “and there are also some other limitations”. Can you add one or a couple of other specific limitations?

-	Consider adding the target audience and coding level required to install and use the package. I assume the library is aimed at hydrologists, geographers, urban planners, etc. In its current format I found the library a bit hard to get started for people that are just getting started.

-	There is a typo in the main Readme file when listing dependency packages. Please, change `matplotlin` for `matplotlib`

-	The caption of Figure 1 in the paper (which also appears in the “Data Model” section of the online documentation) could benefit from a longer caption to better describe the connection between the OOP approach, vertex letters, and variables listed in the boxes. The current figure is fine for people with background in hydrology and computational geometry, but I'm not sure if the figure is clear to a wider audience.

-	In section 4.3.1. it remains unclear whether there are multiple input files or just one. The first and second sentences contradict. Should it say: “Within these configuration files…”?
- There is an empty Readme file in `/pyflowline-main/data/susquehanna/readme.md`. Not sure if this was intentional.

2. Installation instructions: Instructions and dependencies are clearly stated. However, I had a hard time trying to install the package and making it work. Installation was resolved by creating a separate conda environment and then installing the package.

```
conda create --name pyflowline_test
conda install -c conda-forge pyflowline
```

3. Example usage: Authors provided a real-world example, but editing all the file paths for the user’s local installation seems overwhelming for a simple example. It would be nice to have a simpler way to set the paths and run the example from the Jupyter Notebook, since this is probably the first thing that users unfamiliar with the package will try to do to become more familiar with it.

4. Functionality documentation: Great

5. Automated tests: Good

6. Community guidelines: There are guidelines pointing users to the FAQ page, developers contact information, and encouragement to submit a GitHub issue if something did not work. I personally made use of the latter option and the authors responded promptly with clear instructions.

### Software paper

1. Summary: There is a clear description of the high-level functionality of the library. It may be good to provide a definition for “structured” and “unstructured” meshes in layman’s terms for non-technical readers. Perhaps the authors can add a figure illustrating both of these concepts and the perils of using structured rectangular meshes to represent river and streamflow networks. I wonder if the authors can explain some of these concepts using the concepts of vector and raster layers, which I think most people with basic training in geographic information systems will be able to understand and follow. I like the addition of a glossary in the online documentation, and it may be good to add the concepts of structured and unstructured meshes to that section.

2. A statement of need: The statement of need is clear and well-written.

3. State of the field: This seems to be the only software that can generate river networks on structured and unstructured meshes (so if I understand this correctly, the package is basically mesh-independent).

4. Quality of writing: The paper well written. The main website needs some clarification (see my comments above)

5. References: The list of references is complete.
-	The reference by `Engwirda, D., & Liao` is missing the year and the title is all in upper case letters. Please, adopt a consistent style across all references. According to Zenodo the citation should be: `Engwirda, Darren, & Liao, Chang. (2021, October 9). 'Unified' Laguerre-Power Meshes for Coupled Earth System Modelling. 29th International Meshing Roundtable (IMR), Virtual Conference. https://doi.org/10.5281/zenodo.5558988`
-	The reference by Liao et al. is missing the year. It seems that this article was published in 2023 according go the DOI.
