+++ pyflowline/mesh/dggrid/create_dggrid_mesh.py
@@ -118,9 +118,9 @@ def convert_dggrid_mesh_to_pyflowline_mesh(sFilename_dggrid_mesh, sFilename_mesh
     pDataset = pDriver_geojson.CreateDataSource(sFilename_mesh_pyflowline)
     pLayer = pDataset.CreateLayer('cell', pSpatial_reference_gcs, ogr.wkbPolygon)
     # Add one attribute
-    pLayer.CreateField(ogr.FieldDefn('id', ogr.OFTInteger64)) #long type for high resolution
-    pLayer.CreateField(ogr.FieldDefn('lon', ogr.OFTReal)) #long type for high resolution
-    pLayer.CreateField(ogr.FieldDefn('lat', ogr.OFTReal)) #long type for high resolution
+    pLayer.CreateField(ogr.FieldDefn('cellid', ogr.OFTInteger64)) #long type for high resolution
+    pLayer.CreateField(ogr.FieldDefn('longitude', ogr.OFTReal)) #long type for high resolution
+    pLayer.CreateField(ogr.FieldDefn('latitude', ogr.OFTReal)) #long type for high resolution
     pArea_field = ogr.FieldDefn('area', ogr.OFTReal)
     pArea_field.SetWidth(20)
     pArea_field.SetPrecision(2)
@@ -166,9 +166,9 @@ def convert_dggrid_mesh_to_pyflowline_mesh(sFilename_dggrid_mesh, sFilename_mesh
         pPolygon = ogr.Geometry(ogr.wkbPolygon)
         pPolygon.AddGeometry(ring)
         pFeature.SetGeometry(pPolygon)
-        pFeature.SetField("id", int(lCellID) )
-        pFeature.SetField("lon", dLongitude_center )
-        pFeature.SetField("lat", dLatitude_center )
+        pFeature.SetField("cellid", int(lCellID) )
+        pFeature.SetField("longitude", dLongitude_center )
+        pFeature.SetField("latitude", dLatitude_center )
         pFeature.SetField("area", dArea )
       
         pLayer.CreateFeature(pFeature)
