+++ examples/susquehanna/run_simulation_mpas.py
@@ -27,7 +27,8 @@ else:
 #setup case information
 #===================================
 iCase_index = 1
-iFlag_visualization = 0
+iFlag_simulation = 0
+iFlag_visualization = 1
 sMesh = 'mpas'
 sDate='20230701'
 
@@ -38,29 +39,33 @@ oPyflowline = pyflowline_read_model_configuration_file(sFilename_configuration_i
 oPyflowline.aBasin[0].dLatitude_outlet_degree=39.462000
 oPyflowline.aBasin[0].dLongitude_outlet_degree=-76.009300
 
-oPyflowline.setup()
+#oPyflowline.setup()
 if iFlag_visualization ==1:
-    oPyflowline.plot(sVariable_in = 'flowline_filter', sFilename_output_in = 'filter_flowline.png'  )
+    #oPyflowline.plot(sVariable_in = 'flowline_filter', sFilename_output_in = 'filter_flowline.png'  )
     pass
 
-oPyflowline.flowline_simplification()
+if iFlag_simulation == 1:
+    #oPyflowline.flowline_simplification()
+    pass
 
 if iFlag_visualization == 1:
      
     aExtent_meander = [-76.5,-76.2, 41.6,41.9] 
-    oPyflowline.plot( sVariable_in='flowline_simplified' , sFilename_output_in = 'flowline_simplified.png' ) 
+    #oPyflowline.plot( sVariable_in='flowline_simplified' , sFilename_output_in = 'flowline_simplified.png' ) 
 
-    oPyflowline.plot( sVariable_in='flowline_simplified' , sFilename_output_in = 'flowline_simplified_zoom.png', aExtent_in =aExtent_meander ) 
+    #oPyflowline.plot( sVariable_in='flowline_simplified' , sFilename_output_in = 'flowline_simplified_zoom.png', aExtent_in =aExtent_meander ) 
 
     pass
 
-aCell = oPyflowline.mesh_generation()
+if iFlag_simulation == 1:
+    aCell = oPyflowline.mesh_generation()
 
 if iFlag_visualization == 1:
     oPyflowline.plot( sVariable_in='mesh', sFilename_output_in = 'mesh.png' ) 
     pass
 
-oPyflowline.reconstruct_topological_relationship(aCell)
+if iFlag_simulation == 1:
+    oPyflowline.reconstruct_topological_relationship(aCell)
 
 if iFlag_visualization == 1:
     oPyflowline.plot(  sVariable_in='overlap', sFilename_output_in = 'mesh_w_flowline.png',)
