### 2023-07-29 488f882
remove the shapely dependency and update visual
