+++ pyflowline/formats/read_mesh.py
@@ -1,9 +1,9 @@
 import os
 import numpy as np
 from osgeo import ogr, gdal
-from shapely.wkt import loads
+#from shapely.wkt import loads
 from pyflowline.formats.convert_coordinates import convert_gcs_coordinates_to_cell
-
+from pyflowline.external.pyearth.gis.gdal.gdal_functions import get_geometry_coords
 def read_mesh_json(iMesh_type_in, sFilename_mesh_in):
     """
     convert a shpefile to json format.
@@ -35,9 +35,10 @@ def read_mesh_json(iMesh_type_in, sFilename_mesh_in):
     #we also need to spatial reference
     for pFeature_mesh in pLayer_mesh:
         pGeometry_mesh = pFeature_mesh.GetGeometryRef()        
-        dummy0 = loads( pGeometry_mesh.ExportToWkt() )
-        aCoords_gcs = dummy0.exterior.coords
-        aCoords_gcs= np.array(aCoords_gcs)       
+        #dummy0 = loads( pGeometry_mesh.ExportToWkt() )
+        #aCoords_gcs = dummy0.exterior.coords
+        #aCoords_gcs= np.array(aCoords_gcs)
+        aCoords_gcs = get_geometry_coords(pGeometry_mesh)       
         lCellID = pFeature_mesh.GetField("cellid")
         dLon = pFeature_mesh.GetField("longitude")
         dLat = pFeature_mesh.GetField("latitude")        
@@ -92,9 +93,10 @@ def read_mesh_json_w_topology(iMesh_type_in, sFilename_mesh_in):
     #we also need to spatial reference
     for pFeature_mesh in pLayer_mesh:
         pGeometry_mesh = pFeature_mesh.GetGeometryRef()        
-        dummy0 = loads( pGeometry_mesh.ExportToWkt() )
-        aCoords_gcs = dummy0.exterior.coords
-        aCoords_gcs= np.array(aCoords_gcs)       
+        #dummy0 = loads( pGeometry_mesh.ExportToWkt() )
+        #aCoords_gcs = dummy0.exterior.coords
+        #aCoords_gcs= np.array(aCoords_gcs)      
+        aCoords_gcs = get_geometry_coords(pGeometry_mesh)    
         lCellID = pFeature_mesh.GetField("cellid")
         dLon = pFeature_mesh.GetField("longitude")
         dLat = pFeature_mesh.GetField("latitude")        
