### 2023-08-02 #5446 @changliao1025
Hi, @andres-patrignani ,
Thank you for taking the time to write a review of our work. I appreciate your feedback. Below is the full response to your comments.

**General checks reply**

- Thanks for the comment.
- Thanks for the comment.
- Thanks for the comment.

**Functionality reply**
- Thanks for the comment. Indeed creating a new conda environment is a better choice to avoid package conflict. We revised the document for this section.
- Thanks for the comment.
- Thanks for the comment.

**Documentation reply**
- Thanks for the comment. We address your comments one by one below.
- Thanks for the comment. We added the readthedocs link under the About section.
- Thanks for the comment. We moved the quickstart under the title.
- Thanks for the suggestion. We accepted this suggestion in the readthedocs.
- Thanks for the suggestion. We added an example

`For example, existing stream-burning methods always treat the vector river networks as a binary mask and cannot describe the topology near river confluences and meanders.`
- Thanks for the suggestion. Indeed, PyFlowline is a tool/model for computational hydrologists. We added some descriptions in the document to explain some required skillsets to use this tool. We added into the quickstart

`Installing and running PyFlowline requires some basic knowledge of the Python ecosystem.
Besides, configuring a PyFlowline simulation requires some knowledge of Geographic Information System (GIS) and computational hydrology.`

- We fixed the typo.
- Thanks for the comment. We added more details in both figures to explain the graphics.
- Thanks for the comment. The model uses two configuration files. We fixed this error and updated the document.
It was designed to provide a basic description of the data folder. We added some content to it.
- That is a good suggestion. We update that in the document.
- Thanks for the comment on the example. Indeed, setting up the model requires some effort. This is due to two reasons (1) the model was designed to run across scales, so there is a file structure; (2) we try to simplify the configuration files so they contain both paths and parameters; (3) PyFlowline is a core component in HexWatershed, and the configuration is designed to fit both models. As a result, some keywords are not used by PyFlowline. We might improve this part in future development.
- Thanks for the comment.
- Thanks for the comment.
- Thanks for the comment.

**Software paper reply**
- Thanks for the comment. We add a sentence in the paper to define both terms.

`In PyFlowline, we define structured meshes (e.g., lat-lon, raster files with projections, and hexagon) as those with fixed cell sizes and shapes and unstructured meshes as those with variable cell sizes and shapes.`

We also add them to the glossary.
- Thanks for the comment.
- Thanks for the comment.
- Thanks for the comment.
- Thanks for the comment. We fixed the reference issue.
