+++ docs/source/readme.rst
@@ -10,7 +10,7 @@ PyFlowline is a mesh-independent river network generator for hydrologic models.
 
 River networks are landscape features typically represented using vector layers. However, most hydrologic models rely on regular grids to discretize the spatial domain and cannot directly ingest vector features into the model. As a result, hydrologic models usually implement a so-called stream-burning process to convert the vector-based river network into a mesh-based river network. 
 
-However, all the existing stream-burning methods only support the structured meshes and there are also some other limitations. For example, existing stream-burning methods always treat the vector river networks as a binary mask and cannot descrive the topology near river confluences and meanders. 
+However, all the existing stream-burning methods only support the structured meshes and there are also some other limitations. For example, existing stream-burning methods always treat the vector river networks as a binary mask and cannot describe the topology near river confluences and meanders. 
 
 PyFlowline solves this issue by using a mesh-independent approach that intersects the vector river network and mesh to reconstruct the conceptual river network.
 
@@ -30,6 +30,13 @@ As a result, if a spatially-distributed hydrologic model uses the unstructured m
 
 To close this gap, PyFlowline was developed using a mesh-independent approach. At its core, PyFlowline uses the intersection between the vector river network and mesh to reconstruct the conceptual river network.
 
+
+*****************
+Target audience
+*****************
+
+PyFlowline is an advanced modeling tool for hydrologists and hydrologic modelers. 
+Users of PyFlowline should be familiar with basic concepts in Geographic Information System (GIS), including vector and raster data, coordinate systems, and projections.
 
 *****************
 Important notice
