### 2023-08-02 #5446 @changliao1025
I want to add that we recently dropped the 'shapely' dependency after we found alternative GDAL APIs.
