### 2023-08-03 1c09b83
fix a bug that include addi neighbor for bound
