+++ pyflowline/mesh/mpas/create_mpas_mesh.py
@@ -225,15 +225,14 @@ def create_mpas_mesh(iFlag_global_in,
         #now setup the neighbor information
         pmpas.aNeighbor=aNeighborIndex
         pmpas.nNeighbor=len(aNeighborIndex)
-        if pmpas.nNeighbor != pmpas.nVertex:
-            #this cell is next to the ocean boundary
+        if pmpas.nNeighbor != pmpas.nVertex:  #this cell is next to the ocean boundary            
             pmpas.nNeighbor_land = pmpas.nNeighbor 
             pmpas.nNeighbor_ocean = pmpas.nVertex - pmpas.nNeighbor
             pmpas.aNeighbor_land=aNeighborIndex
             pmpas.nNeighbor_land=len(aNeighborIndex)
             print(lCellID)
             print('Warning: nNeighbor != nVertex')
-        else:
+        else: #this cell is not at the the land-ocean mask coastal line
             pmpas.nNeighbor_land = pmpas.nNeighbor 
             pmpas.nNeighbor_ocean = 0
             pmpas.aNeighbor_land = aNeighborIndex        
@@ -242,13 +241,14 @@ def create_mpas_mesh(iFlag_global_in,
         aDistance=list()
         for i in range(pmpas.nNeighbor):
             #find shared edge
-            lEdgeID= aEdgeIndex[i]              
+            lEdgeID= aEdgeIndex[i]
             lIndex = lEdgeID-1
             dDistance = aDcEdge[lIndex]
             aDistance.append(dDistance)
             pass
 
-        pmpas.aNeighbor_distance = aDistance
+        #this contains all the original mpas neighbor distance
+        pmpas.aNeighbor_distance = aDistance 
         aList.append(pmpas)
         return aList
 
@@ -303,7 +303,7 @@ def create_mpas_mesh(iFlag_global_in,
                     else:
                         pass
                 else:
-                    pass      
+                    pass
 
                 #call fuction to add the cell
                 aMpas = add_cell_into_list(aMpas, i, lCellID, dArea, dElevation_mean, dElevation_profile0, aCoords )
@@ -321,8 +321,6 @@ def create_mpas_mesh(iFlag_global_in,
 
                     pLayer.CreateFeature(pFeature)
             
-            
-
     else:
         iFlag_remove_ice = 1
         for i in range(ncell):
@@ -412,7 +410,7 @@ def create_mpas_mesh(iFlag_global_in,
         aMpas_middle = list()
         aMpas_out = list()
         ncell = len(aMpas)
-        #generate the list of cell ID
+        #generate the list of cell ID that are already certain
         aCellID  = list()
         for i in range(ncell):
             pCell = aMpas[i]
@@ -420,33 +418,40 @@ def create_mpas_mesh(iFlag_global_in,
             aCellID.append(lCellID)
 
         #first update neighbor information because some cell should have vitual land neighbor (not present in the mesh)
+        #this operation does not increase the number of cells, but it update the neighbor information
+        #specifically, it divided the land neighbor into two parts: land and virtual land
         for i in range(ncell):
             pCell = aMpas[i]
-            aNeighbor_land = pCell.aNeighbor_land
+            aNeighbor_land = pCell.aNeighbor_land   #including both holes and maps land cutoff by boundary
             nNeighbor_land = pCell.nNeighbor
             aNeighbor_land_update = list()
             aNeighbor_land_virtual = list()
             nNeighbor_land_update = 0 
-            for j in range(nNeighbor_land):
+            for j in range(nNeighbor_land): #loop all land neighbors
                 lNeighbor = int(aNeighbor_land[j])
                 if lNeighbor in aCellID:
                     nNeighbor_land_update = nNeighbor_land_update + 1 
                     aNeighbor_land_update.append(lNeighbor)
                 else:
[…]
