+++ pyflowline/classes/mpas.py
@@ -78,6 +78,7 @@ class pympas(pycell):
     
     aNeighbor_ocean=None #the global ID of all neighbors
     aNeighbor_distance = None
+   
 
     def __init__(self, dLon, dLat, aEdge, aVertex):
         """
