### 2023-08-03 333bde4
docs: edit description of flow network categories in paper
