+++ pyflowline/mesh/square/create_square_mesh.py
@@ -9,6 +9,12 @@ import numpy as np
 from pyflowline.formats.convert_coordinates import convert_gcs_coordinates_to_cell
 from pyflowline.external.pyearth.gis.gdal.gdal_functions import  reproject_coordinates_batch
 
+def index_to_row_col(index, num_columns):
+    index -= 1  # Adjust for 1-based indexing
+    row = index // num_columns + 1
+    col = index % num_columns + 1
+    return row, col
+
 def create_square_mesh(dX_left_in, dY_bot_in,
                         dResolution_meter_in,
                         ncolumn_in, nrow_in,
@@ -64,27 +70,89 @@ def create_square_mesh(dX_left_in, dY_bot_in,
     ybottom = dY_bot_in
     yspacing = dResolution_meter_in
 
-    lCellID = 1 
+    def add_cell_into_list(aList, lCellID, iRow, iColumn, dLongitude_center, dLatitude_center, aCoords ):          
+    
+        pSquare = convert_gcs_coordinates_to_cell(2, dLongitude_center, dLatitude_center, aCoords)
+        pSquare.lCellID = lCellID
+        dArea = pSquare.calculate_cell_area()
+        pSquare.calculate_edge_length()                
+        #build topoloy
+        aNeighbor=list()
+        aNeighbor_distance=list()
+        #lCellID_center = lCellID
+        #counter-clock wise direction to add the neighbor
+        if iRow > 1:#under
+            iRow_dummy = iRow - 1
+            if iColumn > 1:
+                iColumn_dummy = iColumn - 1
+                lCellID2 = (iRow_dummy-1) * ncolumn_in + iColumn_dummy #lCellID0 - nrow_in
+                aNeighbor.append(lCellID2)
+                
+            lCellID0 =  (iRow_dummy-1) * ncolumn_in + iColumn
+            aNeighbor.append(lCellID0)                    
+        if iColumn  < ncolumn_in  : #right
+            iColumn_dummy = iColumn + 1
+            if iRow > 1:
+                iRow_dummy = iRow - 1
+                lCellID7 = (iRow_dummy-1) * ncolumn_in + iColumn_dummy# lCellID5 -1
+                aNeighbor.append(lCellID7) 
+            lCellID5 = (iRow-1) * ncolumn_in + iColumn_dummy #nrow_in * iColumn + iRow 
+            aNeighbor.append(lCellID5)                       
+        if iRow < nrow_in:#top
+            iRow_dummy = iRow + 1
+            if iColumn < ncolumn_in:
+                iColumn_dummy = iColumn + 1
+                lCellID6 = (iRow_dummy-1) * ncolumn_in + iColumn_dummy #lCellID3 + nrow_in
+                aNeighbor.append(lCellID6) 
+            lCellID3 = (iRow_dummy-1) * ncolumn_in + iColumn #lCellID_center + 1
+            aNeighbor.append(lCellID3)         
+        
+        if iColumn> 1:#left
+            iColumn_dummy = iColumn - 1
+            if iRow < nrow_in:
+                iRow_dummy = iRow + 1
+                lCellID4 = (iRow_dummy-1) * ncolumn_in + iColumn_dummy #lCellID1 + 1
+                aNeighbor.append(lCellID4)   
+            lCellID1 = (iRow-1) * ncolumn_in + iColumn_dummy #nrow_in * (iColumn-2) + iRow 
+            aNeighbor.append(lCellID1)  
+
+        pSquare.aNeighbor = aNeighbor
+        pSquare.nNeighbor = len(aNeighbor)
+        pSquare.aNeighbor_land= aNeighbor
+        pSquare.nNeighbor_land= pSquare.nNeighbor
+        aList.append(pSquare)
+
[…]
