### 2023-08-04 cf1533e
simplify some code
