+++ pyflowline/mesh/hexagon/create_hexagon_mesh.py
@@ -76,6 +76,109 @@ def create_hexagon_mesh(iFlag_rotation_in,
     #hexagon edge
     dLength_edge = np.sqrt(  2.0 * dArea / (3.0* np.sqrt(3.0))  )
     dLength_half_edge = 0.5 * dLength_edge
+
+    def add_cell_into_list1(aList, lCellID, iRow, iColumn, dLongitude_center, dLatitude_center, aCoords ):
+        pHexagon = convert_gcs_coordinates_to_cell(1, dLongitude_center, dLatitude_center, aCoords)
+        pHexagon.lCellID = lCellID
+        dArea = pHexagon.calculate_cell_area()
+        pHexagon.calculate_edge_length()
+        
+        lCellID_center = lCellID
+        #build topoloy
+        aNeighbor=list()
+        if iColumn > 1:#0
+            lCellID0 = lCellID_center - 1
+            aNeighbor.append(lCellID0)
+        if iRow < nrow_in :#1 and 2
+            if iRow %2 ==0:
+                lCellID1 = ncolumn_in * iRow + iColumn 
+                aNeighbor.append(lCellID1)
+                if iColumn!=ncolumn_in:
+                    lCellID2 = ncolumn_in * iRow + iColumn + 1
+                    aNeighbor.append(lCellID2)
+            else:
+                lCellID2 = ncolumn_in * iRow + iColumn
+                aNeighbor.append(lCellID2)
+                if iColumn != 1:
+                    lCellID1 = ncolumn_in * iRow + iColumn - 1 
+                    aNeighbor.append(lCellID1)
+        if iColumn < ncolumn_in:#3
+            lCellID3 = lCellID_center + 1
+            aNeighbor.append(lCellID3)             
+        if iRow > 1 : #4 and 5
+            if iRow %2 ==1:
+                lCellID4 = ncolumn_in * (iRow-2) + iColumn 
+                aNeighbor.append(lCellID4)
+                if iColumn !=1:
+                    lCellID5 = ncolumn_in * (iRow-2) + iColumn -1
+                    aNeighbor.append(lCellID5)
+            else:
+                lCellID5 = ncolumn_in * (iRow-2) + iColumn 
+                aNeighbor.append(lCellID5)
+                if iColumn!=ncolumn_in:
+                    lCellID4 = ncolumn_in * (iRow-2) + iColumn + 1
+                    aNeighbor.append(lCellID4)
+        if check_if_duplicates(aNeighbor) == 0:
+            print('error')  
+
+        pHexagon.aNeighbor = aNeighbor
+        pHexagon.nNeighbor = len(aNeighbor)
+        pHexagon.aNeighbor_land= aNeighbor
+        pHexagon.nNeighbor_land= pHexagon.nNeighbor
+        aList.append(pHexagon)  
+        return aList, dArea
+
+
+    def add_cell_into_list2(aList, lCellID, iRow, iColumn, dLongitude_center, dLatitude_center, aCoords ):
+        pHexagon = convert_gcs_coordinates_to_cell(1, dLongitude_center, dLatitude_center, dummy1)
+        pHexagon.lCellID = lCellID
+        dArea = pHexagon.calculate_cell_area()
+        pHexagon.dArea = dArea
+        pHexagon.calculate_edge_length() 
+        
+        lCellID_center = lCellID                
+        aNeighbor=list()
+        if iRow > 1:#0
+            lCellID0 = lCellID_center - 1
+            aNeighbor.append(lCellID0)
+        if iColumn> 1:#1 ans 2
+            if iColumn %2 ==0:
+                lCellID1 = nrow_in * (iColumn-2) + iRow 
+                aNeighbor.append(lCellID1)
[…]
