### 2023-08-04 4b80985
simplify some code
