### 2023-08-04 2c02943
finish hexagon boundary and hole dev
