+++ pyflowline/mesh/hexagon/create_hexagon_mesh.py
@@ -10,6 +10,12 @@ from pyflowline.formats.convert_coordinates import convert_gcs_coordinates_to_ce
 from pyflowline.algorithms.auxiliary.find_index_in_list import check_if_duplicates
 from pyflowline.external.pyearth.gis.gdal.gdal_functions import reproject_coordinates_batch
 
+def index_to_row_col(index, num_columns):
+    index -= 1  # Adjust for 1-based indexing
+    row = index // num_columns + 1
+    col = index % num_columns + 1
+    return row, col
+
 def create_hexagon_mesh(iFlag_rotation_in, 
         dX_left_in, 
         dY_bot_in, 
@@ -86,38 +92,68 @@ def create_hexagon_mesh(iFlag_rotation_in,
         lCellID_center = lCellID
         #build topoloy
         aNeighbor=list()
-        if iColumn > 1:#0
-            lCellID0 = lCellID_center - 1
-            aNeighbor.append(lCellID0)
-        if iRow < nrow_in :#1 and 2
-            if iRow %2 ==0:
-                lCellID1 = ncolumn_in * iRow + iColumn 
+        if iRow > 1 : #1 and 2
+            if iRow %2 ==1:
+                if iColumn !=1:
+                    iRow_dummy = iRow - 1
+                    iColumn_dummy = iColumn - 1
+                    lCellID1 = (iRow_dummy-1) * ncolumn_in + iColumn_dummy  
+                    aNeighbor.append(lCellID1)
+                    
+                iRow_dummy = iRow - 1
+                iColumn_dummy = iColumn 
+                lCellID2 = (iRow_dummy-1) * ncolumn_in + iColumn_dummy
+                aNeighbor.append(lCellID2)
+                
+            else:
+                iRow_dummy = iRow - 1
+                iColumn_dummy = iColumn 
+                lCellID1 = (iRow_dummy-1) * ncolumn_in + iColumn_dummy
                 aNeighbor.append(lCellID1)
                 if iColumn!=ncolumn_in:
-                    lCellID2 = ncolumn_in * iRow + iColumn + 1
+                    iRow_dummy = iRow - 1
+                    iColumn_dummy = iColumn + 1
+                    lCellID2 = (iRow_dummy-1) * ncolumn_in + iColumn_dummy
+                    
                     aNeighbor.append(lCellID2)
-            else:
-                lCellID2 = ncolumn_in * iRow + iColumn
-                aNeighbor.append(lCellID2)
-                if iColumn != 1:
-                    lCellID1 = ncolumn_in * iRow + iColumn - 1 
-                    aNeighbor.append(lCellID1)
+
         if iColumn < ncolumn_in:#3
-            lCellID3 = lCellID_center + 1
-            aNeighbor.append(lCellID3)             
-        if iRow > 1 : #4 and 5
-            if iRow %2 ==1:
-                lCellID4 = ncolumn_in * (iRow-2) + iColumn 
-                aNeighbor.append(lCellID4)
-                if iColumn !=1:
-                    lCellID5 = ncolumn_in * (iRow-2) + iColumn -1
-                    aNeighbor.append(lCellID5)
-            else:
-                lCellID5 = ncolumn_in * (iRow-2) + iColumn 
-                aNeighbor.append(lCellID5)
+            iRow_dummy = iRow 
+            iColumn_dummy = iColumn + 1
+            lCellID3 = (iRow_dummy-1) * ncolumn_in + iColumn_dummy
+            aNeighbor.append(lCellID3)  
+
+        
+        if iRow < nrow_in :#4 and 5
+            if iRow %2 ==0:
                 if iColumn!=ncolumn_in:
-                    lCellID4 = ncolumn_in * (iRow-2) + iColumn + 1
[…]
