### 2023-08-04 b8800d4
finish hexagon boundary and hole dev
