### 2023-08-08 f271223
fix neighbor distance bug
