### 2023-08-09 #5446 @observingClouds
Dear @changliao1025,
I went through the reviews and responses and found that a few issues remain, some of which seem already addressed but not closed:

- https://github.com/changliao1025/pyflowline/issues/165 (seems addressed with changes in the documentation)
- https://github.com/changliao1025/pyflowline/issues/164 (still needs to be addressed, examples should work more straight forward; example files like [these](https://github.com/changliao1025/pyflowline/blob/18209d24dadb498a4e9155a62e56d50dea5ea55a/examples/susquehanna/pyflowline_susquehanna_latlon.json#L36-L39) could be also uploaded to an external permanent storage like zenodo if the files are too big for GitHub or it should be mentioned explicitly if some of them are not used by pyflowline at all to reduce potential confusion; paths that point to files within the GitHub repository could be make relative paths in e.g. the notebook to allow direct execution of the example notebook)
- https://github.com/changliao1025/pyflowline/issues/137 (seems to be addressed)
- https://github.com/changliao1025/pyflowline/issues/136 (seems to be addressed)
> I was able to successfully execute run_simulation_mpas.py in the examples/Susquehanna directory. However, the saved image of the filter_flowline.png has the same issues as decribed above for plotting in the Jupyter Notebook - the file saves but the network is not illustrated in the image file (saved image file shown at the bottom of this review). When I load the corresponding .geojson file into QGIS the network shows up just fine. So, I suspect the same dependency issues described for the Jupyter Notebook are affecting the reproducibility with the .py implementation.

might not yet be addressed.

Please close those issues that have already been addressed.

The execution of the example notebook was most challenging for the reviewers especially because some results could only be reproduced by writing additional code, e.g. by using geopandas to plot and evaluate the results. Because the plotting is a substantial part of the evaluation of this software package it is crucial to make this functionality work out of the box (e.g. by making data paths relative to the notebooks location). I would therefore like you and your co-authors to address these outstanding comments, before I kindly ask @smchartrand and @andres-patrignani to check if all the comments have been addressed appropriately and especially check if the example notebook is easy to follow.

Thank you.
