+++ pyflowline/classes/pycase.py
@@ -102,6 +102,7 @@ class flowlinecase(object):
     iFlag_intersect = 1
     iFlag_rotation=0
     iFlag_break_by_distance = 0  #if the distance between two vertice are far,
+    iResolution_index = 10
     iFlag_dggrid = 0
     nOutlet = 1 #by default , there shoule ne only one ouelet
     dResolution_degree=0.0
@@ -121,6 +122,7 @@ class flowlinecase(object):
     sRegion=''
     sModel=''
     sMesh_type ='hexagon'
+    sDggrid_type = 'ISEA3H'
 
     sCase=''
     sDate=''
@@ -239,6 +241,16 @@ class flowlinecase(object):
             self.iFlag_dggrid = int(aConfig_in['iFlag_dggrid'])
         else:
             self.iFlag_dggrid=0
+
+        if 'iResolution_index' in aConfig_in:
+            self.iResolution_index = int(aConfig_in['iResolution_index'])
+        else:
+            self.iResolution_index=10
+        
+        if 'sDggrid_type' in aConfig_in:
+            self.sDggrid_type = aConfig_in['sDggrid_type']
+        else:
+            self.sDggrid_type='ISEA3H'
 
         if 'nOutlet' in aConfig_in:
             self.nOutlet = int(aConfig_in['nOutlet'])
@@ -770,10 +782,11 @@ class flowlinecase(object):
 
                                     aDggrid = create_dggrid_mesh(iFlag_global,
                                                                      iFlag_save_mesh,
-                                                                     dResolution_meter,
                                                                      sFilename_mesh,
                                                                      sWorkspace_output,
+                                                                     iResolution_index_in= self.iResolution_index,
                                                                      iFlag_antarctic_in=iFlag_antarctic_in,
+                                                                     sDggrid_type_in = self.sDggrid_type,
                                                                      sFilename_boundary_in = self.sFilename_mesh_boundary)
                                                                      
                                     pass
@@ -782,7 +795,9 @@ class flowlinecase(object):
                                                                      iFlag_save_mesh,
                                                                      dResolution_meter,
                                                                      sFilename_mesh,
-                                                                     sWorkspace_output)
+                                                                     sWorkspace_output,
+                                                                      iResolution_index_in= self.iResolution_index,
+                                                                          sDggrid_type_in = self.sDggrid_type)
                                     pass
 
 
@@ -807,7 +822,7 @@ class flowlinecase(object):
                                     print('Unsupported mesh type?')
                                 return
         else:
-            #read mesh?
+            #read mesh? this function is not completed
             iMesh_type = self.iMesh_type
             aCell_out = read_mesh_json_w_topology(iMesh_type, self.sFilename_mesh)
             pass
