### 2023-08-23 f9f4f1e
add set parameter function
