### 2023-08-23 48c5430
add set parameter function
