+++ pyflowline/algorithms/merge/merge_flowline.py
@@ -11,7 +11,8 @@ else:
 
 lID = 0
 
-def merge_flowline(aFlowline_in, aVertex_in, 
+def merge_flowline(aFlowline_in, 
+                   aVertex_in, 
     pVertex_outlet_in, 
         aIndex_headwater_in,
             aIndex_middle_in, 
