+++ pyflowline/algorithms/auxiliary/find_vertex_in_list.py
@@ -1,4 +1,14 @@
+
+import importlib
 import numpy as np
+iFlag_cython = importlib.util.find_spec("cython") 
+if iFlag_cython is not None:
+    from pyflowline.external.tinyr.tinyr.tinyr import RTree
+    iFlag_use_rtree = 1
+else:
+    iFlag_use_rtree =0
+    pass
+
 def find_vertex_in_list(aVertex_in, pVertex_in, dThreshold_in=1.0E-6):
     """[find the index of a vertex in a list]
 
@@ -14,21 +24,60 @@ def find_vertex_in_list(aVertex_in, pVertex_in, dThreshold_in=1.0E-6):
     lIndex= -1
     nVertex= len(aVertex_in)
 
-    if nVertex > 0 :
-        for i in np.arange( nVertex):
-            pVertex = aVertex_in[i]
+    if iFlag_use_rtree == 1:
+
+        #can we use rtree here?
+        #index_vertex = rtree.index.Index()
+        index_vertex = RTree(max_cap=5, min_cap=2)
+        for i in range(nVertex):
+            lID = i 
+            x=aVertex_in[i].dLongitude_degree
+            y=aVertex_in[i].dLatitude_degree   
+            left= x - 1E-5
+            right= x + 1E-5
+            bottom= y-1E-5
+            top=    y+1E-5
+            pBound= (left, bottom, right, top)
+            index_vertex.insert(lID, pBound)  # 
+
+        #now the new vertex
+        x=pVertex_in.dLongitude_degree
+        y=pVertex_in.dLatitude_degree   
+        left = x - 1E-5
+        right = x + 1E-5
+        bottom = y-1E-5
+        top =    y+1E-5
+        pBound= (left, bottom, right, top)
+        aIntersect = list(index_vertex.search(pBound))
+
+        for k in aIntersect:
+            pVertex = aVertex_in[k]
             dDistance = pVertex.calculate_distance(pVertex_in)
             #if pVertex == pVertex_in:
             if dDistance < dThreshold_in:
                 iFlag_exist = 1      
-                lIndex = i 
+                lIndex = k
                 break                
             else:
                 pass
-
-        pass        
-        
-    else:
         pass
+    else:
+
+        if nVertex > 0 :
+            for i in np.arange( nVertex):
+                pVertex = aVertex_in[i]
+                dDistance = pVertex.calculate_distance(pVertex_in)
+                #if pVertex == pVertex_in:
+                if dDistance < dThreshold_in:
+                    iFlag_exist = 1      
+                    lIndex = i 
+                    break                
+                else:
+                    pass
+
+            pass        
+
+        else:
+            pass
     
     return iFlag_exist, lIndex
