+++ pyflowline/algorithms/intersect/intersect_flowline_with_mesh.py
@@ -1,5 +1,6 @@
 
 import os
+import importlib
 import numpy as np
 from osgeo import ogr, osr
 #from shapely.wkt import loads
@@ -8,8 +9,15 @@ from pyflowline.formats.convert_coordinates import convert_gcs_coordinates_to_ce
 from pyflowline.formats.convert_coordinates import convert_gcs_coordinates_to_flowline
 from pyflowline.external.pyearth.gis.gdal.gdal_functions import get_geometry_coords
 
-def intersect_flowline_with_mesh(iMesh_type_in, sFilename_mesh_in, sFilename_flowline_in, sFilename_output_in):
+iFlag_cython = importlib.util.find_spec("cython") 
+if iFlag_cython is not None:
+    from pyflowline.external.tinyr import tinyr
+    iFlag_use_rtree = 1
+else:
+    iFlag_use_rtree =0
+    pass
 
+def intersect_flowline_with_mesh(iMesh_type_in, sFilename_mesh_in, sFilename_flowline_in, sFilename_output_in):
     if  os.path.exists(sFilename_mesh_in) and  os.path.exists(sFilename_flowline_in) : 
         pass
     else:
@@ -23,14 +31,13 @@ def intersect_flowline_with_mesh(iMesh_type_in, sFilename_mesh_in, sFilename_flo
     pDriver_geojson = ogr.GetDriverByName( "GeoJSON")
     aCell=list()
     aCell_intersect=list()
-   
-    pDataset_mesh = pDriver_geojson.Open(sFilename_mesh_in, 0)
-    pDataset_flowline = pDriver_geojson.Open(sFilename_flowline_in, 0)   
 
+    pDataset_mesh = pDriver_geojson.Open(sFilename_mesh_in, 0)
     pLayer_mesh = pDataset_mesh.GetLayer(0)
     pSpatial_reference_mesh = pLayer_mesh.GetSpatialRef()
     nfeature_mesh = pLayer_mesh.GetFeatureCount()
 
+    pDataset_flowline = pDriver_geojson.Open(sFilename_flowline_in, 0)   
     pLayer_flowline = pDataset_flowline.GetLayer(0)
     pSpatial_reference_flowline = pLayer_flowline.GetSpatialRef()
     nfeature_flowline = pLayer_flowline.GetFeatureCount()
@@ -42,7 +49,7 @@ def intersect_flowline_with_mesh(iMesh_type_in, sFilename_mesh_in, sFilename_flo
         transform = osr.CoordinateTransformation(pSpatial_reference_mesh, pSpatial_reference_flowline)
     else:
         iFlag_transform = 0
-
+    
     pDataset_out = pDriver_geojson.CreateDataSource(sFilename_output_in)
 
     pLayerOut = pDataset_out.CreateLayer('flowline', pSpatial_reference_flowline, ogr.wkbMultiLineString)
@@ -51,126 +58,249 @@ def intersect_flowline_with_mesh(iMesh_type_in, sFilename_mesh_in, sFilename_flo
     pLayerOut.CreateField(ogr.FieldDefn('stream_segment', ogr.OFTInteger)) #long type for high resolution
     pLayerOut.CreateField(ogr.FieldDefn('stream_order', ogr.OFTInteger)) #long type for high resolution
     pLayerDefn = pLayerOut.GetLayerDefn()
-    pFeatureOut = ogr.Feature(pLayerDefn)       
+    pFeatureOut = ogr.Feature(pLayerDefn)   
+
     lID_flowline = 0          
-    aFlowline_intersect_all=list()   
-    for pFeature_mesh in pLayer_mesh:       
-        pGeometry_mesh = pFeature_mesh.GetGeometryRef()        
-        #dummy0 = loads( pGeometry_mesh.ExportToWkt() )
[…]
