+++ pyflowline/algorithms/simplification/remove_small_river.py
@@ -1,5 +1,14 @@
 from pyflowline.algorithms.auxiliary.check_head_water import check_head_water
 def remove_small_river(aFlowline_in, dThreshold_in):
+    """Remove small river that meet the threshold and headwater requirement, also dam flowline are reserved
+
+    Args:
+        aFlowline_in (_type_): _description_
+        dThreshold_in (_type_): _description_
+
+    Returns:
+        List: Theortically, the flowline should be ordered from outlet to headwater
+    """
     nFlowline = len(aFlowline_in)
     aFlowline_out=list()        
     if nFlowline == 1:
@@ -17,8 +26,7 @@ def remove_small_river(aFlowline_in, dThreshold_in):
                 aFlowline_out.append(pFlowline)
                 lID = lID +1       
             else:
-
-                if check_head_water(aFlowline_in, pVertex_start)==1:
+                if pFlowline.iStream_order == 1:
                     if dLength > dThreshold_in :
                         pFlowline.lIndex = lID
                         aFlowline_out.append(pFlowline)
@@ -26,11 +34,25 @@ def remove_small_river(aFlowline_in, dThreshold_in):
                         pass
                     else:
                         pass
-                else:        
+                    pass
+                else: #this one might be not used     
                     pFlowline.lIndex = lID
                     aFlowline_out.append(pFlowline)
                     lID = lID +1       
-                    pass            
+                    pass        
+                    #if check_head_water(aFlowline_in, pVertex_start)==1:
+                    #    if dLength > dThreshold_in :
+                    #        pFlowline.lIndex = lID
+                    #        aFlowline_out.append(pFlowline)
+                    #        lID = lID + 1 
+                    #        pass
+                    #    else:
+                    #        pass
+                    #else:        
+                    #    pFlowline.lIndex = lID
+                    #    aFlowline_out.append(pFlowline)
+                    #    lID = lID +1       
+                    #    pass            
             
             pass    
 
