+++ pyflowline/algorithms/index/define_stream_order.py
@@ -39,12 +39,8 @@ def define_stream_order(aFlowline_in, aConfluence_in):
     """
     nFlowline = len(aFlowline_in)
     nSegment = nFlowline
-    aFlowline_out = list()
-
-  
-    iMethod = 1
-    
-    
+    aFlowline_out = list()  
+    iMethod = 1       
     if nFlowline == 0 :
         print ('data incomplete')
     else:       
