+++ pyflowline/classes/basin.py
@@ -511,27 +511,31 @@ class pybasin(object):
             aFlowline_basin_simplified = correct_flowline_direction(aFlowline_basin_simplified,  pVertex_outlet )
             nFlowline_after = len(aFlowline_basin_simplified)
             ptimer.stop()
+            sys.stdout.flush()
+            pVertex_outlet = aFlowline_basin_simplified[0].pVertex_end
+            self.pVertex_outlet = pVertex_outlet
+            if self.iFlag_debug ==1:
+                sFilename_out = 'flowline_direction_before_intersect.geojson'
+                sFilename_out = os.path.join(sWorkspace_output_basin, sFilename_out)
+                export_flowline_to_geojson( aFlowline_basin_simplified,  sFilename_out)
         except:
-            print(nFlowline_before)
+            print('Error in flow direction correction')
             
-        sys.stdout.flush()
-        pVertex_outlet = aFlowline_basin_simplified[0].pVertex_end
-        self.pVertex_outlet = pVertex_outlet
-        if self.iFlag_debug ==1:
-            sFilename_out = 'flowline_direction_before_intersect.geojson'
-            sFilename_out = os.path.join(sWorkspace_output_basin, sFilename_out)
-            export_flowline_to_geojson( aFlowline_basin_simplified,  sFilename_out)
+        
         
         #step 4: remove loops
-        print('Basin ',  self.sBasinID, 'started loop removal')
-        ptimer.start()
-        aFlowline_basin_simplified = remove_flowline_loop(aFlowline_basin_simplified)   
-        ptimer.stop() 
-        sys.stdout.flush()
-        if self.iFlag_debug ==1:
-            sFilename_out = 'flowline_loop_before_intersect.geojson'
-            sFilename_out = os.path.join(sWorkspace_output_basin, sFilename_out)
-            export_flowline_to_geojson( aFlowline_basin_simplified, sFilename_out)
+        try:
+            print('Basin ',  self.sBasinID, 'started loop removal')
+            ptimer.start()
+            aFlowline_basin_simplified = remove_flowline_loop(aFlowline_basin_simplified)   
+            ptimer.stop() 
+            sys.stdout.flush()
+            if self.iFlag_debug ==1:
+                sFilename_out = 'flowline_loop_before_intersect.geojson'
+                sFilename_out = os.path.join(sWorkspace_output_basin, sFilename_out)
+                export_flowline_to_geojson( aFlowline_basin_simplified, sFilename_out)
+        except:
+            print('Error in loop removal')
         
         
         #at this stage, the stream order information is not available, but could be easily rebuild for speed up the small river removal algorithm
@@ -669,7 +673,7 @@ class pybasin(object):
         Returns:
             list [pyflowline]: A list of intersected cells
         """
-        print('Start topology reconstruction:',  self.sBasinID)
+        print('Basin ',  self.sBasinID, 'Start topology reconstruction')
         
         ptimer = pytimer()
         
@@ -677,17 +681,20 @@ class pybasin(object):
         sFilename_flowline_in = self.sFilename_flowline_simplified
         sFilename_flowline_intersect_out = self.sFilename_flowline_intersect
 
-        ptimer.start()
-        aCell, aCell_intersect_basin, aFlowline_intersect_all = intersect_flowline_with_mesh(iMesh_type, sFilename_mesh, \
-            sFilename_flowline_in, sFilename_flowline_intersect_out)
+        try:
+            print('Basin ',  self.sBasinID, 'Start flowline and mesh intersection')
+            ptimer.start()
[…]
