+++ pyflowline/classes/square.py
@@ -72,7 +72,7 @@ class pysquare(pycell):
     aNeighbor_land=None #the global ID of all neighbors
     aNeighbor_ocean=None #the global ID of all neighbors
     aNeighbor_distance = None
-   
+    pBound=None
 
     def __init__(self, dLon, dLat, aEdge, aVertex):    
         """
@@ -102,9 +102,23 @@ class pysquare(pycell):
             self.iStream_order_burned=-1
             self.iStream_segment_burned=-1
             self.dElevation_mean=-9999.0
+            self.calculate_cell_bound() #bound for rtree 
             pass
         pass
     
+    def calculate_cell_bound(self):
+        dLat_min = 90
+        dLat_max = -90
+        dLon_min = 180
+        dLon_max = -180
+        for i in range(self.nVertex):
+            dLon_max = np.max( [dLon_max, self.aVertex[i].dLongitude_degree] )
+            dLon_min = np.min( [dLon_min, self.aVertex[i].dLongitude_degree] )
+            dLat_max = np.max( [dLat_max, self.aVertex[i].dLatitude_degree] )
+            dLat_min = np.min( [dLat_min, self.aVertex[i].dLatitude_degree] )
+        
+        self.pBound = (dLon_min, dLat_min, dLon_max, dLat_max)
+        return self.pBound
     def has_this_edge(self, pEdge_in):
         """
         Check whether the square contains an edge
