+++ pyflowline/classes/tin.py
@@ -54,7 +54,7 @@ class pytin(pycell):
     nNeighbor=-1
 
     aNeighbor_distance = None
-  
+    pBound=None
 
     def __init__(self, aEdge,aVertex, dLon, dLat):       
         nEdge = len(aEdge)
@@ -71,9 +71,22 @@ class pytin(pycell):
             pVertex['lon'] =self.dLongitude_center
             pVertex['lat'] =self.dLatitude_center           
             self.pVertex_center = pyvertex(pVertex)
+            self.calculate_cell_bound() #bound for rtree 
             pass
         pass    
-   
+    def calculate_cell_bound(self):
+        dLat_min = 90
+        dLat_max = -90
+        dLon_min = 180
+        dLon_max = -180
+        for i in range(self.nVertex):
+            dLon_max = np.max( [dLon_max, self.aVertex[i].dLongitude_degree] )
+            dLon_min = np.min( [dLon_min, self.aVertex[i].dLongitude_degree] )
+            dLat_max = np.max( [dLat_max, self.aVertex[i].dLatitude_degree] )
+            dLat_min = np.min( [dLat_min, self.aVertex[i].dLatitude_degree] )
+        
+        self.pBound = (dLon_min, dLat_min, dLon_max, dLat_max)
+        return self.pBound
     def has_this_edge(self, pEdge_in):
         iFlag_found = 0
         for pEdge in self.aEdge:

[…]
