+++ pyflowline/classes/basin.py
@@ -732,10 +732,11 @@ class pybasin(object):
             print('Error in split_flowline_to_edge.')
 
         try:
-            print('Basin ',  self.sBasinID, 'Start split flowline to edge')          
+            print('Basin ',  self.sBasinID, 'Start remove duplicate flowline')          
             ptimer.start()
             aFlowline_basin_conceptual = remove_duplicate_flowline(aFlowline_basin_conceptual)
             ptimer.stop()
+            sys.stdout.flush()
             if self.iFlag_debug ==1:
                 sFilename_out = 'flowline_edge_remove_duplicate_flowline.geojson'
                 sFilename_out = os.path.join(sWorkspace_output_basin, sFilename_out)
@@ -748,6 +749,7 @@ class pybasin(object):
             ptimer.start()
             aFlowline_basin_conceptual = correct_flowline_direction(aFlowline_basin_conceptual,  pVertex_outlet )
             ptimer.stop()
+            sys.stdout.flush()
             if self.iFlag_debug ==1:
                 sFilename_out = 'flowline_edge_correct_flowline_direction.geojson'
                 sFilename_out = os.path.join(sWorkspace_output_basin, sFilename_out)
@@ -760,6 +762,7 @@ class pybasin(object):
             ptimer.start()
             aFlowline_basin_conceptual = remove_flowline_loop(aFlowline_basin_conceptual )  
             ptimer.stop()
+            sys.stdout.flush()
             if self.iFlag_debug ==1:
                 sFilename_out = 'flowline_edge_remove_flowline_loop.geojson'
                 sFilename_out = os.path.join(sWorkspace_output_basin, sFilename_out)
@@ -775,6 +778,7 @@ class pybasin(object):
             aVertex, lIndex_outlet, aIndex_headwater,aIndex_middle, aIndex_confluence, aConnectivity, pVertex_outlet\
                 = find_flowline_confluence(aFlowline_basin_conceptual,  pVertex_outlet)
             ptimer.stop()
+            sys.stdout.flush()
             if self.iFlag_debug ==1:
                 sFilename_out = 'flowline_vertex_with_confluence_after_intersect.geojson'
                 sFilename_out = os.path.join(sWorkspace_output_basin, sFilename_out)
@@ -788,6 +792,7 @@ class pybasin(object):
             ptimer.start()
             aFlowline_basin_conceptual = merge_flowline( aFlowline_basin_conceptual,aVertex, pVertex_outlet, aIndex_headwater,aIndex_middle, aIndex_confluence  )     
             ptimer.stop()
+            sys.stdout.flush()
         except:
             print('Error in merge_flowline.')
 
@@ -799,6 +804,7 @@ class pybasin(object):
             aVertex, lIndex_outlet, aIndex_headwater,aIndex_middle, aIndex_confluence, aConnectivity, pVertex_outlet\
                 = find_flowline_confluence(aFlowline_basin_conceptual,  pVertex_outlet)   
             ptimer.stop()
+            sys.stdout.flush()
         except:
             print('Error in find_flowline_confluence.')
              
@@ -819,6 +825,7 @@ class pybasin(object):
         ptimer.start()
         aFlowline_basin_conceptual = define_stream_topology(aFlowline_basin_conceptual, aConfluence_basin_conceptual)
         ptimer.stop()
+        sys.stdout.flush()
 
         aFlowline_basin_conceptual, aStream_order = define_stream_order(aFlowline_basin_conceptual, aConfluence_basin_conceptual)
 
