+++ pyflowline/external/pyearth/gis/gdal/gdal_functions.py
@@ -415,7 +415,7 @@ def gdal_read_geotiff_file(sFilename_in):
         pBand = None      
         pBand = None
 
-        return aData_out, dPixelWidth, dOriginX, dOriginY, nrow, ncolumn, dMissing_value, pGeotransform, pProjection,  pSpatial_reference
+        return aData_out, dPixelWidth, pPixelHeight, dOriginX, dOriginY, nrow, ncolumn, dMissing_value, pGeotransform, pProjection,  pSpatial_reference
 
 def Google_MetersPerPixel( zoomLevel ):  
    
