+++ pyflowline/classes/pycase.py
@@ -305,8 +305,8 @@ class flowlinecase(object):
 
             if self.iFlag_mesh_boundary==1:
                 if not os.path.isfile(self.sFilename_mesh_boundary ):
-                    print("The mesh boundary file does not exist!")
-                    exit()
+                    print("The mesh boundary file does not exist, you should update this parameter before running the model!")
+                    #exit()
                 pass
 
         if 'sFilename_spatial_reference' in aConfig_in:
@@ -320,8 +320,8 @@ class flowlinecase(object):
             if 'sFilename_dggrid' in aConfig_in:
                 self.sFilename_dggrid = aConfig_in['sFilename_dggrid']
                 if not os.path.isfile(self.sFilename_dggrid ):
-                    print("The dggrid binary file does not exist!")
-                    exit()
+                    print("The dggrid binary file does not exist, you need to update this parameter before running the model!")
+                    #exit()
                 pass
 
             else:
@@ -427,7 +427,8 @@ class flowlinecase(object):
                     pass
                 else:
                     print('This basin configuration file does not exist: ', self.sFilename_basins )
-                    exit()
+                    print('Please update this parameter before running the model!')
+                    #exit()
                 with open(self.sFilename_basins) as json_file:
                     dummy_data = json.load(json_file)
                     for i in range(self.nOutlet):
@@ -1014,6 +1015,7 @@ class flowlinecase(object):
                 sType_input = type(dValue)
                 if sType_default == sType_input:      
                     setattr(self, sVariable_in, dValue)
+                    self.update_model()
                     pass   
                 else:
                     print('Incorrect data type for the input value: ' + sVariable_in)
@@ -1021,6 +1023,9 @@ class flowlinecase(object):
             else:
                 print("This model parameter is unknown, please check the full parameter list in the documentation: " + sVariable_in)
                 return False
+            
+            
+
         else:
             #this mean the variable is in the basin object
             for pBasin in self.aBasin:
@@ -1037,7 +1042,28 @@ class flowlinecase(object):
                     print("This model parameter is unknown, please check the full parameter list in the documentation: " + sVariable_in)
                     return False
 
-    
+    def update_model(self):
+        """This function should be called after the model parameter is changed
+        """
+        #update subbasin
+        if self.iFlag_flowline==1:            
+            if os.path.isfile(self.sFilename_basins):
+                pass
+            else:                
+                print("Error: the basin file does not exist: " + self.sFilename_basins)
+            
+            self.aBasin.clear()
+
+            with open(self.sFilename_basins) as json_file:
+                dummy_data = json.load(json_file)
+                for i in range(self.nOutlet):
+                    sBasin =  "{:08d}".format(i+1)
+                    dummy_basin = dummy_data[i]
+                    dummy_basin['sWorkspace_output_basin'] = str(Path(self.sWorkspace_output) / sBasin )
+                    pBasin = pybasin(dummy_basin)
+                    self.aBasin.append(pBasin)
+        
+
     def run(self):
         """
         Run the flowlinecase simulation
