### 2023-09-28 #5446 @changliao1025
Hi,  @observingClouds , sorry for the late updates. We recently made many upgrades to the pyflowline model, including the spatial index algorithm, to speed up the algorithm.

As for the I/O part, we added two additional functions (see example in the notebook), to set and change the model parameters.
These improvements are designed to ease how users interact with the model parameters. With these two functions, the users can change the parameters without directly editing the JSON files.

Besides, we simplify the visualization part using GeoPandas.
The notebook is only intended to serve as an illustration of the workflow. Due to the computational nature of the model, it is recommended to use the model in the Python environment.
