+++ pyflowline/classes/_visual_basin.py
@@ -24,6 +24,7 @@ def basin_plot(self,
                iFlag_title_in=None,
                iFlag_colorbar_in=None,
                iFlag_scientific_notation_colorbar_in=None,
+               iFlag_openstreetmap_in = None,
                dData_min_in = None,
                dData_max_in = None,
                sVariable_in=None,
@@ -104,7 +105,8 @@ def basin_plot(self,
                             map_multiple_vector_data(aFiletype_in,
                                              aFilename_in,  
                                              iFlag_title_in=iFlag_title_in,                         
-                                             iFont_size_in=iFont_size_in,                 
+                                             iFont_size_in=iFont_size_in,          
+                                             iFlag_openstreetmap_in=iFlag_openstreetmap_in,       
                                              sFilename_output_in=sFilename_output_in,
                                              sTitle_in= 'Flow direction with observation',
                                              aFlag_thickness_in=  [1, 0, 0],
