+++ pyflowline/mesh/square/create_square_mesh.py
@@ -136,6 +136,8 @@ def create_square_mesh(dX_left_in, dY_bot_in,
     #(x1,y1)-----(x2,y2)
     #...............
     aSquare = list()
+    aSquare_dict = dict()
+    lCellIndex = 0
     for iRow in range(1, nrow_in+1):    
         for iColumn in range(1, ncolumn_in+1):        
             #global cell id for the mesh
@@ -191,7 +193,7 @@ def create_square_mesh(dX_left_in, dY_bot_in,
                 
                 
             if ( iFlag == True ):
-                aSquare, dArea = add_cell_into_list(aSquare, lCellID, iRow, iColumn, dLongitude_center,dLatitude_center, dummy1 ) 
+                aSquare, dArea = add_cell_into_list(aSquare, lCellID, iRow, iColumn, dLongitude_center,dLatitude_center, aCoords ) 
         
                 #save feature
                 pFeature.SetGeometry(pPolygon)
@@ -201,12 +203,16 @@ def create_square_mesh(dX_left_in, dY_bot_in,
                 pFeature.SetField("area", dArea )
                 pLayer.CreateFeature(pFeature)
 
+                #add to dictionary
+                aSquare_dict[lCellID] = lCellIndex
+                lCellIndex = lCellIndex + 1
+
                 pass
 
     
+    iFlag_fill_hole = 0
     aSquare_out = list()
-    aSquare_middle = list()
- 
+    aSquare_middle = list() 
     ncell = len(aSquare)
     aCellID  = list()
     for i in range(ncell):
@@ -214,155 +220,170 @@ def create_square_mesh(dX_left_in, dY_bot_in,
         lCellID = pCell.lCellID
         aCellID.append(lCellID)
 
-    for i in range(ncell):
-        pCell = aSquare[i]
-        aNeighbor_land = pCell.aNeighbor_land   #including both holes and maps land cutoff by boundary
-        nNeighbor_land = pCell.nNeighbor
-        aNeighbor_land_update = list()
-        aNeighbor_land_virtual = list()
-        nNeighbor_land_update = 0 
-        for j in range(nNeighbor_land): #loop all land neighbors
-            lNeighbor = int(aNeighbor_land[j])
-            if lNeighbor in aCellID:
-                nNeighbor_land_update = nNeighbor_land_update + 1 
-                aNeighbor_land_update.append(lNeighbor)
-            else:
-                #a hole or boundary mpas land cell
-                aNeighbor_land_virtual.append(lNeighbor)
-                
-        pCell.nNeighbor= len(aNeighbor_land_update)
-        pCell.aNeighbor = aNeighbor_land_update        
-        pCell.aNeighbor_land = aNeighbor_land_update
-        pCell.nNeighbor_land= len(aNeighbor_land_update)   
-        pCell.aNeighbor_land_virtual = aNeighbor_land_virtual   
-        
-        pCell.nNeighbor_land_virtual = len(aNeighbor_land_virtual)
-        aSquare_middle.append(pCell)
+    if iFlag_fill_hole == 1:       
+        #find virtual land cells
+        for i in range(ncell):
+            pCell = aSquare[i]
+            aNeighbor_land = pCell.aNeighbor_land   #including both holes and maps land cutoff by boundary
+            nNeighbor_land = pCell.nNeighbor
+            aNeighbor_land_update = list()
+            aNeighbor_land_virtual = list()
+            nNeighbor_land_update = 0 
+            for j in range(nNeighbor_land): #loop all land neighbors
+                lNeighbor = int(aNeighbor_land[j])
+                if lNeighbor in aCellID:
+                    nNeighbor_land_update = nNeighbor_land_update + 1 
+                    aNeighbor_land_update.append(lNeighbor)
+                else:
+                    #a hole or boundary mpas land cell
[…]
