+++ pyflowline/mesh/latlon/create_latlon_mesh.py
@@ -56,6 +56,60 @@ def create_latlon_mesh(dLongitude_left_in,
     ybottom = dLatitude_bot_in
     yspacing = dResolution_degree_in
 
+    def add_cell_into_list(aList, lCellID, iRow, iColumn, dLongitude_center, dLatitude_center, aCoords ):          
+    
+        pLatlon = convert_gcs_coordinates_to_cell(2, dLongitude_center, dLatitude_center, aCoords)
+        pLatlon.lCellID = lCellID
+        dArea = pLatlon.calculate_cell_area()
+        pLatlon.calculate_edge_length()                
+        #build topoloy
+        aNeighbor=list()
+        aNeighbor_distance=list()
+        #lCellID_center = lCellID
+        #counter-clock wise direction to add the neighbor
+        if iRow > 1:#under
+            iRow_dummy = iRow - 1
+            if iColumn > 1:
+                iColumn_dummy = iColumn - 1
+                lCellID2 = (iRow_dummy-1) * ncolumn_in + iColumn_dummy #lCellID0 - nrow_in
+                aNeighbor.append(lCellID2)
+                
+            lCellID0 =  (iRow_dummy-1) * ncolumn_in + iColumn
+            aNeighbor.append(lCellID0)                    
+        if iColumn  < ncolumn_in  : #right
+            iColumn_dummy = iColumn + 1
+            if iRow > 1:
+                iRow_dummy = iRow - 1
+                lCellID7 = (iRow_dummy-1) * ncolumn_in + iColumn_dummy# lCellID5 -1
+                aNeighbor.append(lCellID7) 
+            lCellID5 = (iRow-1) * ncolumn_in + iColumn_dummy #nrow_in * iColumn + iRow 
+            aNeighbor.append(lCellID5)                       
+        if iRow < nrow_in:#top
+            iRow_dummy = iRow + 1
+            if iColumn < ncolumn_in:
+                iColumn_dummy = iColumn + 1
+                lCellID6 = (iRow_dummy-1) * ncolumn_in + iColumn_dummy #lCellID3 + nrow_in
+                aNeighbor.append(lCellID6) 
+            lCellID3 = (iRow_dummy-1) * ncolumn_in + iColumn #lCellID_center + 1
+            aNeighbor.append(lCellID3)         
+        
+        if iColumn> 1:#left
+            iColumn_dummy = iColumn - 1
+            if iRow < nrow_in:
+                iRow_dummy = iRow + 1
+                lCellID4 = (iRow_dummy-1) * ncolumn_in + iColumn_dummy #lCellID1 + 1
+                aNeighbor.append(lCellID4)   
+            lCellID1 = (iRow-1) * ncolumn_in + iColumn_dummy #nrow_in * (iColumn-2) + iRow 
+            aNeighbor.append(lCellID1)  
+
+        pLatlon.aNeighbor = aNeighbor
+        pLatlon.nNeighbor = len(aNeighbor)
+        pLatlon.aNeighbor_land= aNeighbor
+        pLatlon.nNeighbor_land= pLatlon.nNeighbor
+        aList.append(pLatlon)
+
+        return aList, dArea
+
     #change the order because mpas uses counter-clock wise to store the vertices
     #we will also start from the lower-left corner, and then go to the right and then go up
     #so the final index will be like this
@@ -68,6 +122,9 @@ def create_latlon_mesh(dLongitude_left_in,
     #(x1,y1)-----(x2,y2)
     #...............
     aLatlon = list()
+    #create a dictionary for searching using unsorted map harsh table method
+    aLatlon_dict = dict()
+    lCellIndex = 0
     for iRow in range(1, nrow_in+1):
[…]
