+++ pyflowline/mesh/mpas/create_mpas_mesh.py
@@ -197,6 +197,8 @@ def create_mpas_mesh(iFlag_global_in,
     aBed_elevation_profile = bed_elevation_profile0[:]  #elevation    
     ncell = len(aIndexToCellID) 
     aMpas = list()
+    aMpas_dict = dict()
+    lCellIndex=0
 
     #add a mpas cell into a list
     def add_cell_into_list(aList, i, lCellID, dArea,dElevation_mean,dElevation_profile0, aCoords  ):
@@ -306,7 +308,8 @@ def create_mpas_mesh(iFlag_global_in,
 
                 #call fuction to add the cell
                 aMpas = add_cell_into_list(aMpas, i, lCellID, dArea, dElevation_mean, dElevation_profile0, aCoords )
-                
+                aMpas_dict[lCellID] = lCellIndex
+                lCellIndex = lCellIndex + 1
                 #save mesh cell
                 if iFlag_save_mesh_in ==1:                
                     pFeature.SetGeometry(pPolygon)
@@ -379,9 +382,10 @@ def create_mpas_mesh(iFlag_global_in,
                 #call fuction to add the cell               
                 
                 aMpas = add_cell_into_list(aMpas, i, lCellID, dArea, dElevation_mean, dElevation_profile0, aCoords )
-                
+                aMpas_dict[lCellID] = lCellIndex
+                lCellIndex = lCellIndex + 1
                 #save mesh cell
-                if iFlag_save_mesh_in ==1:                
+                if iFlag_save_mesh_in == 1:                
                     pFeature.SetGeometry(pPolygon)
                     pFeature.SetField("cellid", int(lCellID) )
                     pFeature.SetField("longitude", dLon )
@@ -397,38 +401,25 @@ def create_mpas_mesh(iFlag_global_in,
     #for maps we need to clean some cell because they were not actually in the domain
     #besides, we need to add some smal holes back
     #to do this, we need two steps.
-
+    
 
     if iFlag_global_in == 1:
         aMpas_out = aMpas
     else:
-        iFlag_fill_hole = 1
-        aMpas_middle = list()
+        iFlag_fill_hole = 0        
         aMpas_out = list()
         ncell = len(aMpas)
-        #generate the list of cell ID that are already certain
-        aCellID  = list()
-        for i in range(ncell):
-            pCell = aMpas[i]
-            lCellID = pCell.lCellID
-            aCellID.append(lCellID)
-
-        
+        #generate the list of cell ID that are already certain        
         if iFlag_fill_hole == 1:   
             #first update neighbor information because some cell should have vitual land neighbor (not present in the mesh)
             #this operation does not increase the number of cells, but it update the neighbor information
             #specifically, it divided the land neighbor into two parts: land and virtual land         
-            for i in range(ncell):
-                pCell = aMpas[i]
+            for pCell in aMpas:               
                 aNeighbor_land = pCell.aNeighbor_land   #including both holes and maps land cutoff by boundary
-                nNeighbor_land = pCell.nNeighbor
                 aNeighbor_land_update = list()
                 aNeighbor_land_virtual = list()
-                nNeighbor_land_update = 0 
-                for j in range(nNeighbor_land): #loop all land neighbors
-                    lNeighbor = int(aNeighbor_land[j])
-                    if lNeighbor in aCellID:
-                        nNeighbor_land_update = nNeighbor_land_update + 1 
+                for lNeighbor in aNeighbor_land: #loop all land neighbors                    
+                    if lNeighbor in aMpas_dict:                        
                         aNeighbor_land_update.append(lNeighbor)
                     else:
                         #a hole or boundary mpas land cell
@@ -438,7 +429,8 @@ def create_mpas_mesh(iFlag_global_in,
[…]
