+++ pyflowline/mesh/mpas/create_mpas_mesh.py
@@ -188,8 +188,8 @@ def create_mpas_mesh(iFlag_global_in,
     aVertexOnCell = verticesOnCell0[:]
     aVertexOnEdge0 = verticesOnEdge0[:]    
     aIndexToCellID = indexToCellID0[:]
-    aIndexToEdgeID = indexToEdgeID0[:]
-    aIndexToVertexID = indexToVertexID0[:]
+    #aIndexToEdgeID = indexToEdgeID0[:]
+    #aIndexToVertexID = indexToVertexID0[:]
     aBed_elevation = bed_elevation0[:]
     aIce_thickness = ice_thickness0[:]
     aCellArea = areaCell0[:]
@@ -201,9 +201,11 @@ def create_mpas_mesh(iFlag_global_in,
     lCellIndex=0
 
     #add a mpas cell into a list
-    def add_cell_into_list(aList, i, lCellID, dArea,dElevation_mean,dElevation_profile0, aCoords  ):
-        dLat = convert_360_to_180 (aLatitudeCell[i])
+    def add_cell_into_list(aList, i, lCellID, dArea, dElevation_mean, dElevation_profile0, aCoords  ):
         dLon = convert_360_to_180 (aLongitudeCell[i])
+        dLat =  (aLatitudeCell[i])        
+        if dLon > 100:
+            print('Warning: longitude > 100')
         #vertex
         aCellOnCellIndex = np.array(aCellsOnCell[i,:])
         aEdgesOnCellIndex = np.array(aEdgesOnCell[i,:])
@@ -240,9 +242,9 @@ def create_mpas_mesh(iFlag_global_in,
             pmpas.nNeighbor_land=len(aNeighborIndex)
 
         aDistance=list()
-        for i in range(pmpas.nNeighbor):
+        for j in range(pmpas.nNeighbor):
             #find shared edge
-            lEdgeID= aEdgeIndex[i]
+            lEdgeID= aEdgeIndex[j]
             lIndex = lEdgeID-1
             dDistance = aDcEdge[lIndex]
             aDistance.append(dDistance)
@@ -258,8 +260,9 @@ def create_mpas_mesh(iFlag_global_in,
         #if it is antarctic, we dont need the boundary
         for i in range(ncell):
             #center
-            dLat = convert_360_to_180 (aLatitudeCell[i])
             dLon = convert_360_to_180 (aLongitudeCell[i])
+            dLat =  (aLatitudeCell[i])
+            
             aVertexOnCellIndex = np.array(aVertexOnCell[i,:])
             dummy0 = np.where(aVertexOnCellIndex > 0)
             aVertexIndex = aVertexOnCellIndex[dummy0]
@@ -327,8 +330,7 @@ def create_mpas_mesh(iFlag_global_in,
         iFlag_remove_ice = 1
         for i in range(ncell):
             #center
-            dLat = convert_360_to_180 (aLatitudeCell[i])
-            dLon = convert_360_to_180 (aLongitudeCell[i])
+                  
             #vertex
             aVertexOnCellIndex = np.array(aVertexOnCell[i,:])
             dummy0 = np.where(aVertexOnCellIndex > 0)
@@ -337,7 +339,7 @@ def create_mpas_mesh(iFlag_global_in,
             aLatVertex = aLatitudeVertex[aVertexIndex-1]
             nVertex = len(aLonVertex)
             #first check if it is within the boundary
-            iFlag = False
+            
             ring = ogr.Geometry(ogr.wkbLinearRing)
             aCoords = np.full((nVertex,2), -9999.0, dtype=float)
             for j in range(nVertex):
@@ -358,11 +360,17 @@ def create_mpas_mesh(iFlag_global_in,
             if pPolygon.Within(pBoundary):
                 iFlag = True
             else:
[…]
