+++ pyflowline/classes/basin.py
@@ -18,6 +18,13 @@ from pyflowline.formats.convert_flowline_to_geojson import convert_flowline_to_g
 from pyflowline.formats.export_flowline import export_flowline_to_geojson
 from pyflowline.formats.export_vertex import export_vertex_to_geojson
 from pyflowline.external.pyearth.toolbox.reader.text_reader_string import text_reader_string
+iFlag_cython = importlib.util.find_spec("cython") 
+if iFlag_cython is not None:
+    from pyflowline.external.tinyr.tinyr.tinyr import RTree
+    iFlag_use_rtree = 1
+else:
+    iFlag_use_rtree =0
+    pass
 
 from pyflowline.algorithms.split.find_flowline_vertex import find_flowline_vertex
 from pyflowline.algorithms.split.find_flowline_confluence import find_flowline_confluence
@@ -126,6 +133,7 @@ class pybasin(object):
     aFlowline_basin_filtered=None
     aFlowline_basin_simplified=None
     aFlowline_basin_conceptual=None    
+    aFlowline_basin_edge = None
     pVertex_outlet=None
     aConfluence_basin_simplified= None
     aConfluence_basin_conceptual= None
@@ -150,6 +158,9 @@ class pybasin(object):
 
     sFilename_variable_polygon=''
     sFilename_variable_polyline=''
+
+    pRTree_flowline = None
+    pRTree_edge = None
 
     
     iFlag_visual = importlib.util.find_spec("cartopy") 
@@ -329,7 +340,7 @@ class pybasin(object):
         self.sFilename_flowline_conceptual_kml = os.path.join(str(self.sWorkspace_output_basin ),'flowline_conceptual.kml')
         return
         
-    def flowline_simplification(self):
+    def basin_flowline_simplification(self):
         """
         Run the basin flowline simplification
 
@@ -463,10 +474,10 @@ class pybasin(object):
         if self.iFlag_debug ==1:
             sFilename_out = 'flowline_before_intersect.geojson'
             sFilename_out = os.path.join(sWorkspace_output_basin, sFilename_out)            
-            self.export_flowline(aFlowline_basin_filtered, sFilename_out)
+            self.basin_export_flowline(aFlowline_basin_filtered, sFilename_out)
         #calculate length
         self.aFlowline_basin_filtered = aFlowline_basin_filtered
-        self.dLength_flowline_filtered = self.calculate_flowline_length(aFlowline_basin_filtered)
+        self.dLength_flowline_filtered = self.basin_calculate_flowline_length(aFlowline_basin_filtered)
 
         #assign vertex id
         
@@ -629,7 +640,7 @@ class pybasin(object):
             print('Basin ',  self.sBasinID, 'started confluence definition')
             ptimer.start()
             aVertex_confluence = aVertex[aIndex_confluence]
-            aConfluence_basin_simplified = self.build_confluence(aFlowline_basin_simplified, aVertex_confluence) 
+            aConfluence_basin_simplified = self.basin_build_confluence(aFlowline_basin_simplified, aVertex_confluence) 
             ptimer.stop()
 
         #change added a new function to build stream topology
@@ -665,7 +676,7 @@ class pybasin(object):
         sys.stdout.flush()
         return aFlowline_basin_simplified
 
-    def reconstruct_topological_relationship(self, iMesh_type, sFilename_mesh):
+    def basin_reconstruct_topological_relationship(self, iMesh_type, sFilename_mesh):
         """
         Run the basin topologic relationship reconstruction
 
@@ -818,7 +829,7 @@ class pybasin(object):
         aIndex_confluence = np.array(aIndex_confluence)
         if aIndex_confluence.size > 0:        
             aVertex_confluence = aVertex[aIndex_confluence] 
[…]
