### 2023-10-26 #5446 @andres-patrignani
I want to thank the authors for the extensive work to improve the pyflowline package. I was able to successfully:

- re-install the pyflowline package using the provided step by step guidelines in the documentation
- run the new Jupyter Notebook from top to bottom for the example channel network and mesh

Features that I would like to see implemented in this or future versions (I leave this up to the editor and other reviewers):
- I was unable to find detailed guidelines on how users can create/use their own files for creating channel networks and meshes. For instance, if I want to find the channel network for a watershed in my area, what information and files do I need? The example seems to be using a NetCDF file, but is unclear how to get similar files for a different area. If I miss this part in the documentation, please let me know.

Overall I think that the code and the manuscript are in good shape and can be accepted.

Some additional comments for future improvement beyond JOSS:
- It may make more sense to use booleans (True and False) for the "iFlag" variables rather than "1" and "0". This is more of a semantic preference than a technical aspect.
- In the example notebook, consider using a coarse mesh for faster mesh generation (new users may wonder why it takes so long to run some lines). Some cells took several minutes to run.
- The notebook example is still a bit too verbose and will probably benefit from some additional Markdown to create sections  and add some notes on what some of the input parameters mean. Perhaps the authors can gradually improve this part as they provide assistance to new users.
- Building a small library of simple examples could be a great resource for new users aiming at using this tool for their own applications.

I hope that my comments have been useful and I commend the authors for creating this library.

Andres
