+++ conda-recipe/conda_build_config.yaml
@@ -23,7 +23,7 @@ metadata:
   # Package name
   name: pyflowline
   # Package version
-  version: "0.2.6"
+  version: "0.2.7"
   # Package summary
   summary: A mesh-independent river network generator for hydrologic models.
   # Package homepage
