### 2023-10-27 aca36dd
remove middle list
