+++ pyflowline/formats/export_vertex.py
@@ -117,14 +117,18 @@ def export_vertex_to_shapefile(aVertex_in, sFilename_shapefile_out,\
     lID = 0
     for i in range(nVertex):       
         pVertex = aVertex_in[i]
+        pPoint = ogr.Geometry(ogr.wkbPoint)
         if iFlag_projected_in ==1:
-            dummy1= Point( pVertex.dx, pVertex.dy )             
+            #dummy1= Point( pVertex.dx, pVertex.dy )             
+            pPoint.AddPoint(pVertex.dx, pVertex.dy)
             pass
         else:
-            dummy1= Point( pVertex.dLongitude_degree, pVertex.dLatitude_degree ) 
+            #dummy1= Point( pVertex.dLongitude_degree, pVertex.dLatitude_degree ) 
+            pPoint.AddPoint(pVertex.dLongitude_degree, pVertex.dLatitude_degree)
             pass
 
-        pGeometry_out = ogr.CreateGeometryFromWkb(dummy1.wkb)
+        #pGeometry_out = ogr.CreateGeometryFromWkb(dummy1.wkb)
+        pGeometry_out = ogr.CreateGeometryFromWkb(pPoint.ExportToWkb())
         pFeature_out.SetGeometry(pGeometry_out)   
         pFeature_out.SetField("id", lID)
         if iFlag_attribute ==1:
