### 2023-10-27 #5446 @changliao1025
@andres-patrignani Thank you for the thoughtful comments, which helped us a lot in improving our package.

Here are some quick responses to your questions:

Indeed, the mesh generation process, especially unstructured meshes such as the MPAS mesh, has various technical challenges. Currently, pyflowline has built-in support to generate structured meshes, see reference in JAMES paper. But it does not have built-in unstructured mesh generators. Instead, it can read the outputs from these unstructured mesh generators, such as JIGSAW-MPAS by @dengwirda and the newly added DGGRID mesh feature by @sahrk. There are a certain amount of learning curves to using an unstructured mesh generator, considering whether it has good documentation. The MPAS mesh is unique because it involves a range of operations. However, it has many advantages that other meshes cannot provide. That is why we use it as an example to demonstrate the capability of pyflowline to support fully unstructured meshes. In the future, we may add a more straightforward example/notebook to use a different mesh, such as the DGGRID mesh. We have just released such a dataset on Zenodo: [https://zenodo.org/records/8377766], which can be customized for this purpose.

Again, thanks for the suggestions, and we will keep improving it in future versions.
