### 2023-10-31 #5446 @smchartrand
I want to echo the comments by @andres-patrignani, and thank the authors for their extensive efforts to improve pyflowline. All of my previous comments have been addressed (and done so very well), and yesterday I was able to:

- (re)install pyflowline using the detailed documentation and step by step instructions (Ubuntu 22.04 LTS release) , and

- implement an instance of the updated Jupyter Notebook and run each notebook component all the way through for the example provided by the authors.

I have no further suggestions. Congratulations to the authors on an excellent package. Thanks for the opportunity to review.
