### 2023-11-06 71d7ac2
Fix typos in paper.md
