### 2023-11-06 #5446 @observingClouds
@changliao1025 thank you for finishing the post-review tasks. I have a few minor comments:
- [ ] The zenodo package title is "river networks generator", while everywhere else it is called "river network generator". Please adapt accordingly so that all resources match.
- [ ] I found a few typos in the paper and opened a [PR](https://github.com/changliao1025/pyflowline/pull/200). Feel free to merge it or correct those yourself.
- [ ] Please make sure that the capitalisation in your references is preserved. You can preserve capitalisation by placing {} around characters/words in your .bib file. Liao et al. (2023b) is one example where the capitalisation is not preserved. There is also something wrong with the DOI url.
- [ ] Please add the volume/issue information to Liao et al. (2023a).

Please address these points and we should be soon ready for publication.
