+++ paper.bib
@@ -99,8 +99,8 @@ year = {2012}
   year         = 2022,
   publisher    = {Zenodo},
   version      = {0.1.20},
-  doi          = {10.5281/zenodo.6368652},
-  url          = {https://doi.org/10.5281/zenodo.6368652}
+  doi          = {10.5281/zenodo.6109987},
+  url          = {https://doi.org/10.5281/zenodo.6109987}
 }
 
 @article{Sahr:2011,
