+++ paper.bib
@@ -32,7 +32,7 @@ volume = {15},
 number = {2},
 pages = {e2022MS003089},
 keywords = {watershed, land-river-ocean interaction, flow direction, river network, unstructured mesh, graph theory},
-doi = {https://doi.org/10.1029/2022MS003089},
+doi = {10.1029/2022MS003089},
 url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2022MS003089},
 eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2022MS003089},
 note = {e2022MS003089 2022MS003089},
