### 2023-11-06 #5446 @changliao1025
Hi, @observingClouds; all the above issues are fixed now. The version is also bumped to 0.3.4.
Thank you for identifying these issues.
